refactor(library): skills controller (skills series 2/3, wave-4 task 2)
Moves 86 of 127 "skill"-named LibraryScreen methods, byte-for-byte, onto
a new LibrarySkillsController (UI/Library_Modules/library_skills_controller.py).
LibraryScreen keeps one-line delegators under every original name (the one
staticmethod, `_restore_library_skills_scope`, forwards to the controller
class). `library_skill_import_controller.py`/`library_skills_browse_
controller.py` stay untouched per plan; the screen still owns both live
instances, now reached via injected accessors.

Single-vs-split: ONE controller, decided by call-graph connected-components
analysis over the 121 movable candidates (127 raw minus 6 merely-delegate
properties) -- one component of 107 members plus 13 disconnected leaves,
no clean editor/trust vs browse/list seam despite the plan's own hypothesis.

41 exclusions from the raw 127: 6 merely-delegate-to-existing-controller
properties, 27 unbound-fake-self test-bypass, 1 instance-attribute
monkeypatch, 1 module-globals coupling (`save_setting_to_cli_config`), and
6 bare-self-as-identity-argument hazard exclusions -- a new bypass shape
this series adds to the recipe: passing bare `self` to code that compares
it for IDENTITY against the app's real current screen, which a controller
standing in for `self` can never satisfy. One was found by static analysis
(`self.workers.cancel_group(self, ...)`, Textual's own node-identity
filter). Five were found only by running the verification battery after a
first draft moved them: `_library_screen_is_current(self)` (four Import-row
handlers) and an inline `self.app.screen is self` guard
(`_present_library_skills_import_snapshot`) both permanently evaluate to
"stale" once `self` is the controller, silently breaking the entire Skills
import flow. Caught by two real Pilot-driven UI tests plus 8 Tests/Skills
tests (the wave's own fourth-root trap, confirmed live); all five methods
reverted to screen-resident, full-bodied, with named dependency bindings
for their remaining mover callers where any exist.

Screen ratchet: 43179/1311 -> 41247/1311 (pure move, methods unchanged).
Controller ratchet: born-governed, pinned at 3099 lines.

See library_skills_controller.py's own module docstring for the complete
cluster derivation, binding-kind reasoning, and per-exclusion evidence.

Recipe: backlog/docs/library-decomposition-recipe.md. Wave-4 task 2
(skills series 2/3).
