+++ tldw_chatbook/UI/Screens/library_screen.py
@@ -579,6 +579,7 @@ from ..Library_Modules.library_notes_work_session import (
 )
 from ..Library_Modules.library_rag_search_controller import LibraryRagSearchController
 from ..Library_Modules.library_rag_search_state import LibraryRagSearchState
+from ..Library_Modules.library_skills_controller import LibrarySkillsController
 from ..Library_Modules.library_skills_state import (
     LibrarySkillsState,
     skill_state_shim_attr,
@@ -2388,6 +2389,113 @@ class LibraryScreen(BaseAppScreen):
             ),
             refresh_search_rag_panel_state_widgets=(
                 lambda **k: self._refresh_search_rag_panel_state_widgets(**k)
+            ),
+        )
+        self._skills_controller = LibrarySkillsController(
+            self,
+            skills_state_accessor=lambda: self._skills_state,
+            library_skill_import_coordinator_accessor=(
+                lambda: self._library_skill_import_coordinator
+            ),
+            set_library_skill_import_coordinator=(
+                lambda value: setattr(self, "_library_skill_import_coordinator", value)
+            ),
+            library_skills_browse_controller_accessor=(
+                lambda: self._library_skills_browse_controller
+            ),
+            set_library_skills_browse_controller=(
+                lambda value: setattr(self, "_library_skills_browse_controller", value)
+            ),
+            run_library_service_call=(
+                lambda *args, **kwargs: self._run_library_service_call(
+                    *args, **kwargs
+                )
+            ),
+            sanitize_media_field=lambda *a, **k: self._sanitize_media_field(*a, **k),
+            sanitize_note_content=(
+                lambda *a, **k: self._sanitize_note_content(*a, **k)
+            ),
+            refresh_local_source_snapshot=(
+                lambda *a, **k: self._refresh_local_source_snapshot(*a, **k)
+            ),
+            library_entry_route_key=lambda: self._library_entry_route_key(),
+            library_entry_reconcile_is_current=(
+                lambda *a, **k: self._library_entry_reconcile_is_current(*a, **k)
+            ),
+            capture_library_entry_focus=lambda: self._capture_library_entry_focus(),
+            restore_library_entry_focus=(
+                lambda *a, **k: self._restore_library_entry_focus(*a, **k)
+            ),
+            library_selected_row_id_accessor=lambda: self._library_selected_row_id,
+            set_library_selected_row_id=(
+                lambda value: setattr(self, "_library_selected_row_id", value)
+            ),
+            library_snapshot_state_generation_accessor=(
+                lambda: self._library_snapshot_state_generation
+            ),
+            library_entry_reconcile_dirty_accessor=(
+                lambda: self._library_entry_reconcile_dirty
+            ),
+            library_entry_reconcile_pending_accessor=(
+                lambda: self._library_entry_reconcile_pending
+            ),
+            library_canvas_projection_depth_accessor=(
+                lambda: self._library_canvas_projection_depth
+            ),
+            library_canvas_resync_pending_accessor=(
+                lambda: self._library_canvas_resync_pending
+            ),
+            set_library_canvas_resync_pending=(
+                lambda value: setattr(self, "_library_canvas_resync_pending", value)
+            ),
+            library_skills_import_open_accessor=(
+                lambda: self._library_skills_import_open
+            ),
+            library_skills_import_path_accessor=(
+                lambda: self._library_skills_import_path
+            ),
+            library_skills_import_status_accessor=(
+                lambda: self._library_skills_import_status
+            ),
+            library_skills_import_review_name_accessor=(
+                lambda: self._library_skills_import_review_name
+            ),
+            library_skills_import_in_flight_accessor=(
+                lambda: self._library_skills_import_in_flight
+            ),
+            library_skills_import_generation_accessor=(
+                lambda: self._library_skills_import_generation
+            ),
+            set_library_skills_import_generation=(
+                lambda value: setattr(self, "_library_skills_import_generation", value)
+            ),
+            approve_library_skill_trust=(
+                lambda *a, **k: self._approve_library_skill_trust(*a, **k)
+            ),
+            begin_library_skill_save=(
+                lambda *a, **k: self._begin_library_skill_save(*a, **k)
+            ),
+            build_library_skills_state=lambda: self._build_library_skills_state(),
+            call_library_skill_trust_service=(
+                lambda *a, **k: self._call_library_skill_trust_service(*a, **k)
+            ),
+            exit_library_skill_editor_guarded=(
+                lambda *a, **k: self._exit_library_skill_editor_guarded(*a, **k)
+            ),
+            persist_library_skill_editor_mode=(
+                lambda *a, **k: self._persist_library_skill_editor_mode(*a, **k)
+            ),
+            refresh_library_skills_trust_posture=(
+                lambda: self._refresh_library_skills_trust_posture()
+            ),
+            request_library_skills_browse=(
+                lambda *a, **k: self._request_library_skills_browse(*a, **k)
+            ),
+            reset_library_skill_editor_state=(
+                lambda *a, **k: self._reset_library_skill_editor_state(*a, **k)
+            ),
+            start_library_skills_import=(
+                lambda: self._start_library_skills_import()
             ),
         )
         (
@@ -6919,54 +7027,8 @@ class LibraryScreen(BaseAppScreen):
         shell.sync_layout(layout)
         self._library_prompts_reader_layout = layout
 
-    def _sync_library_skills_reader_layout_from_shell(
-        self,
-        priority: Literal["library", "items"] | None = None,
-    ) -> None:
-        """Resolve the settled Skills shell and patch it in place."""
-        try:
-            shell = self.query_one(
-                "#library-skills-reader-shell", LibraryAdaptiveReaderShell
-            )
-        except (NoMatches, QueryError):
-            return
-        width = shell.region.width
-        if width <= 0:
-            return
-        previous = self._library_skills_reader_layout
-        if (
-            previous.reader_width == 0
-            and previous.library_width == 0
-            and previous.items_width == 0
-        ):
-            previous = None
-        if (
-            priority is None
-            and self._library_skills_view == "list"
-            and self._library_skills_reader_preferences.items_open
-        ):
-            items_priority_floor = (
-                2 * PANE_GRIP_WIDTH
-                + LIBRARY_SKILLS_READER_PROFILE.list_min_width
-                + LIBRARY_SKILLS_READER_PROFILE.work_min_width
-            )
-            if width >= items_priority_floor:
-                priority = "items"
-        elif (
-            priority is None
-            and previous is not None
-            and previous.priority_pane == "items"
-        ):
-            previous = dataclasses.replace(previous, priority_pane=None)
-        layout = resolve_adaptive_reader_layout(
-            width,
-            self._library_skills_reader_preferences,
-            LIBRARY_SKILLS_READER_PROFILE,
-            previous=previous,
-            priority=priority,
-        )
-        shell.sync_layout(layout)
-        self._library_skills_reader_layout = layout
+    def _sync_library_skills_reader_layout_from_shell(self, priority: Literal['library', 'items'] | None = None) -> None:
+        return self._skills_controller._sync_library_skills_reader_layout_from_shell(priority)
 
     def _sync_library_conversation_reader_layout_from_shell(
         self,
@@ -7059,25 +7121,8 @@ class LibraryScreen(BaseAppScreen):
             library_config[section_name] = section
         section[key] = value
 
-    def _mirror_library_skills_reader_preference(
-        self,
-        key: Literal["library_open", "items_open"],
-        value: bool,
-    ) -> None:
-        """Mirror one optimistic Skills pane choice into app config."""
-        app_config = getattr(self.app_instance, "app_config", None)
-        if not isinstance(app_config, dict):
-            return
-        library_config = app_config.setdefault("library", {})
-        if not isinstance(library_config, dict):
-            library_config = {}
-            app_config["library"] = library_config
-        section_name = "reader" if key == "library_open" else "skills_reader"
-        section = library_config.setdefault(section_name, {})
-        if not isinstance(section, dict):
-            section = {}
-            library_config[section_name] = section
-        section[key] = value
+    def _mirror_library_skills_reader_preference(self, key: Literal['library_open', 'items_open'], value: bool) -> None:
+        return self._skills_controller._mirror_library_skills_reader_preference(key, value)
 
     def _replace_library_reader_preference(
         self,
@@ -9583,33 +9628,7 @@ class LibraryScreen(BaseAppScreen):
 
     @staticmethod
     def _restore_library_skills_scope(state: Mapping[str, Any]) -> SkillBrowseScope:
-        """Return a dispatch-safe applied Skills scope from screen state."""
-        saved = state.get("library_skills_scope")
-        if isinstance(saved, SkillBrowseScope):
-            raw = dataclasses.asdict(saved)
-        elif type(saved) is dict:
-            raw = saved
-        else:
-            raw = {
-                "query": state.get("library_skills_filter", ""),
-                "sort": state.get("library_skills_sort", "name"),
-                "page": 1,
-            }
-        query = raw.get("query", "")
-        page = raw.get("page", 1)
-        if type(query) is not str:
-            query = ""
-        if type(page) is not int or not 1 <= page <= MAX_SKILL_BROWSE_PAGE:
-            page = 1
-        try:
-            return SkillBrowseScope(
-                query=query,
-                sort=raw.get("sort", "name"),
-                page=page,
-                page_size=DEFAULT_SKILL_BROWSE_PAGE_SIZE,
-            )
-        except (TypeError, ValueError):
-            return SkillBrowseScope()
+        return LibrarySkillsController._restore_library_skills_scope(state)
 
     @staticmethod
     def _restore_library_collections_page(state: Mapping[str, Any]) -> int:
@@ -12880,47 +12899,8 @@ class LibraryScreen(BaseAppScreen):
             return None
         return result if isinstance(result, int) else None
 
-    async def _skills_context_or_none(
-        self, get_context: Any, **kwargs: Any
-    ) -> Mapping[str, Any] | None:
-        """Fetch the local skills context, degrading quietly on failure.
-
-        Runs inside the same ``asyncio.gather`` as the notes/media/
-        conversations/prompts fetch (see ``_list_local_source_snapshot``).
-        Mirrors ``_prompts_count_or_none``: the seam is optional (guarded by
-        ``callable(get_context)`` at the call site), so when it is missing
-        this method is never invoked, and when it *is* present but raises,
-        the failure is swallowed and ``None`` is returned -- the Skills
-        rail row then renders uncounted with an empty context payload
-        rather than surfacing an error or failing the whole snapshot
-        fetch.
-
-        Unlike prompts (whose rail count and exact browse result have
-        separate owners), a single ``get_context`` call here supplies both:
-        the count is derived from its
-        ``available_skills``/``blocked_skills`` lengths by the caller, and
-        the same payload is stashed for a future Skills canvas to render.
-
-        Args:
-            get_context: The bound ``skills_scope_service.get_context``
-                callable to invoke.
-            **kwargs: Forwarded to ``get_context`` (``mode``).
-
-        Returns:
-            The normalized ``get_context`` payload (``available_skills`` +
-            ``blocked_skills``), or ``None`` if the call failed or returned
-            something other than a ``Mapping``.
-        """
-        try:
-            result = await self._run_library_service_call(
-                get_context, isolate_in_worker=True, **kwargs
-            )
-        except Exception:
-            logger.opt(exception=True).warning(
-                "Failed to fetch local skills context; Skills row will show no count."
-            )
-            return None
-        return result if isinstance(result, Mapping) else None
+    async def _skills_context_or_none(self, get_context: Any, **kwargs: Any) -> Mapping[str, Any] | None:
+        return await self._skills_controller._skills_context_or_none(get_context, **kwargs)
 
     async def _study_count_or_none(
         self, count_callable: Any, label: str, **kwargs: Any
@@ -18152,169 +18132,16 @@ class LibraryScreen(BaseAppScreen):
         return values
 
     def _build_library_skill_tool_catalog(self) -> tuple[str, ...]:
-        """Return the existing builtin/local tool names a Skill may restrict."""
-        try:
-            from ...Agents.tool_catalog import BuiltinToolProvider
-
-            names = [entry.name for entry in BuiltinToolProvider().list_catalog()]
-            console_config = getattr(self.app_instance, "app_config", {}).get(
-                "console", {}
-            )
-            local_enabled = bool(
-                isinstance(console_config, Mapping)
-                and coerce_bool_setting(
-                    console_config.get("local_tools_enabled"), False
-                )
-            )
-            if local_enabled:
-                from ...Agents.local_tool_provider import LocalToolProvider
-
-                configured_root = console_config.get("workspace_root")
-                workspace_root = (
-                    Path(str(configured_root)).expanduser()
-                    if configured_root
-                    else Path.cwd()
-                )
-                names.extend(
-                    entry.name
-                    for entry in LocalToolProvider(
-                        workspace_root=workspace_root
-                    ).list_catalog()
-                )
-        except Exception:
-            logger.warning("Library Skill tool catalog is temporarily unavailable.")
-            return ()
-        return tuple(dict.fromkeys(names))
+        return self._skills_controller._build_library_skill_tool_catalog()
 
     def _library_skills_list_canvas_kwargs(self) -> dict[str, Any]:
-        """Return list-only inputs for the retained Skills Items pane."""
-        values = self._library_skills_canvas_kwargs(consume_scroll=False)
-        values.update(
-            state=self._build_library_skills_state(),
-            mode="list",
-            editor_state=None,
-            warnings="",
-            status="",
-            conflict=False,
-            active_review=None,
-            confirming_delete=False,
-            scroll_to_actions=False,
-            skill_path="",
-            import_open=False,
-            import_path="",
-            import_status="",
-            import_review_name="",
-            mutation_in_flight=False,
-            more_actions_open=False,
-            trust_details_open=False,
-            script_access_granted=False,
-        )
-        return values
+        return self._skills_controller._library_skills_list_canvas_kwargs()
 
     def _library_skill_work_pane_kwargs(self) -> dict[str, Any]:
-        """Return the active non-list projection for the Skills Work pane."""
-        values = self._library_skills_canvas_kwargs()
-        if self._library_skills_import_open:
-            values["mode"] = "list"
-        values["state"] = None
-        values["sort_choices_visible"] = False
-        values["reader_mode"] = self._library_skill_reader_mode
-        return values
+        return self._skills_controller._library_skill_work_pane_kwargs()
 
-    def _library_skills_canvas_kwargs(
-        self, *, consume_scroll: bool = True
-    ) -> dict[str, Any]:
-        """Return every compose input for the mounted Skills canvas."""
-        values: dict[str, Any] = {
-            "state": None,
-            "sort_mode": self._library_skills_sort,
-            "filter_value": self._library_skills_filter,
-            "mode": "list",
-            "trust_posture": self._library_skills_trust_posture,
-            "confirming_reset": self._library_skill_trust_confirming_reset,
-            "editor_state": None,
-            "warnings": "",
-            "status": "",
-            "conflict": False,
-            "active_review": None,
-            "is_create": not self._selected_skill_name,
-            "dirty": self._library_skill_dirty,
-            "confirming_delete": self._library_skill_confirming_delete,
-            "scroll_to_actions": False,
-            "skill_path": "",
-            "import_open": False,
-            "import_path": "",
-            "import_status": "",
-            "import_review_name": "",
-            "import_in_flight": False,
-            "import_package_kind": "",
-            "import_recovery_actions": (),
-            "import_retryable": False,
-            "sort_choices_visible": False,
-            "editor_mode": self._library_skill_editor_mode,
-            "tool_catalog": self._library_skill_tool_catalog,
-            "tool_filter": self._library_skill_tool_filter,
-            "mutation_in_flight": self._library_skill_mutation_in_flight,
-            "more_actions_open": self._library_skill_more_actions_open,
-            "trust_details_open": self._library_skill_trust_details_open,
-            "script_access_granted": self._library_skill_script_grant,
-            "detail_notice": "",
-            "detail_retryable": False,
-        }
-        if self._library_skills_view == "editor":
-            editor_state = self._library_skill_editor_state
-            if editor_state is None:
-                values["mode"] = "loading"
-                values["detail_notice"] = (
-                    self._library_skill_detail_error or "Loading skill…"
-                )
-                values["detail_retryable"] = self._library_skill_detail_retryable
-            else:
-                values.update(
-                    {
-                        "mode": "editor",
-                        "editor_state": editor_state,
-                        "warnings": "\n".join(
-                            skill_editor_warning_lines(
-                                live_name=editor_state.name,
-                                trust_status=editor_state.trust_status,
-                                trust_blocked=editor_state.trust_blocked,
-                            )
-                        ),
-                        "status": self._library_skill_status,
-                        "conflict": self._library_skill_conflict,
-                        "active_review": self._library_skill_active_review,
-                        "scroll_to_actions": (
-                            self._consume_library_skill_scroll_pending()
-                            if consume_scroll
-                            else False
-                        ),
-                        "skill_path": self._library_skill_on_disk_path(),
-                    }
-                )
-            return values
-
-        values.update(
-            {
-                "state": self._build_library_skills_state(),
-                "import_open": self._library_skills_import_open,
-                "import_path": self._library_skills_import_path,
-                "import_status": self._library_skills_import_status,
-                "import_review_name": self._library_skills_import_review_name,
-                "import_in_flight": self._library_skills_import_in_flight,
-                "import_package_kind": (
-                    self._library_skill_import_coordinator.snapshot.package_kind
-                ),
-                "import_recovery_actions": (
-                    self._library_skill_import_coordinator.snapshot.recovery_actions
-                ),
-                "import_retryable": (
-                    self._library_skill_import_coordinator.snapshot.retryable
-                ),
-                "sort_choices_visible": self._library_skills_sort_choices_visible,
-            }
-        )
-        return values
+    def _library_skills_canvas_kwargs(self, *, consume_scroll: bool = True) -> dict[str, Any]:
+        return self._skills_controller._library_skills_canvas_kwargs(consume_scroll=consume_scroll)
 
     def _build_library_prompts_state(self):
         """Build the Library prompts canvas's list-view display state.
@@ -18704,113 +18531,14 @@ class LibraryScreen(BaseAppScreen):
             focus_identity=focus_identity,
         )
 
-    def _sync_library_skills_browse_result(
-        self,
-        result: SkillBrowseResult,
-        focus_identity: str | None,
-    ) -> LibraryEntryReconcileResult:
-        """Project one accepted Skills generation and restore stable focus."""
-        controller = self._library_skills_browse_controller
-        if (
-            result.request_token != controller.result.request_token
-            or self._library_selected_row_id != LIBRARY_ROW_BROWSE_SKILLS
-            or self._library_skills_view != "list"
-        ):
-            return LibraryEntryReconcileResult.SUPERSEDED
-
-        self._library_skills_filter = controller.scope.query
-        self._library_skills_sort = controller.scope.sort
-        focused = getattr(self, "focused", None)
-        live_focus_id = getattr(focused, "id", None)
-        if (
-            isinstance(focused, (Button, Input))
-            and isinstance(live_focus_id, str)
-            and live_focus_id.startswith("library-skills-")
-        ):
-            focus_identity = live_focus_id
-        cursor_context = self._library_skills_filter_cursor_context
-        cursor_position = (
-            focused.cursor_position
-            if isinstance(focused, Input) and live_focus_id == LIBRARY_SKILLS_FILTER_ID
-            else cursor_context[1]
-            if cursor_context is not None and cursor_context[0] == result.request_token
-            else None
-        )
-
-        def restore_focus() -> None:
-            if not focus_identity:
-                return
-            if result.status == "loading" and focus_identity in {
-                LIBRARY_SKILLS_PAGE_PREVIOUS_ID,
-                LIBRARY_SKILLS_PAGE_NEXT_ID,
-                LIBRARY_SKILLS_RETRY_ID,
-            }:
-                return
-            if focus_identity in {
-                LIBRARY_SKILLS_PAGE_PREVIOUS_ID,
-                LIBRARY_SKILLS_PAGE_NEXT_ID,
-                LIBRARY_SKILLS_RETRY_ID,
-            }:
-                self._focus_library_skills_page_control(focus_identity)
-                return
-            try:
-                target = self.query_one(f"#{focus_identity}", Widget)
-            except (NoMatches, QueryError):
-                return
-            if not getattr(target, "disabled", False):
-                target.focus()
-                if cursor_position is not None and isinstance(target, Input):
-                    target.cursor_position = cursor_position
-
-        if _sync_library_canvas(
-            self,
-            "skills",
-            then=restore_focus,
-            allow_screen_fallback=False,
-        ):
-            return LibraryEntryReconcileResult.APPLIED
-        return LibraryEntryReconcileResult.FAILED
+    def _sync_library_skills_browse_result(self, result: SkillBrowseResult, focus_identity: str | None) -> LibraryEntryReconcileResult:
+        return self._skills_controller._sync_library_skills_browse_result(result, focus_identity)
 
     def _focus_library_skills_page_control(self, invoked: str) -> None:
-        """Restore pager focus without landing on a disabled control."""
-        opposite = {
-            LIBRARY_SKILLS_PAGE_PREVIOUS_ID: LIBRARY_SKILLS_PAGE_NEXT_ID,
-            LIBRARY_SKILLS_PAGE_NEXT_ID: LIBRARY_SKILLS_PAGE_PREVIOUS_ID,
-            LIBRARY_SKILLS_RETRY_ID: LIBRARY_SKILLS_FILTER_ID,
-        }[invoked]
-        for control_id in (invoked, opposite, LIBRARY_SKILLS_FILTER_ID):
-            try:
-                control = self.query_one(f"#{control_id}", Widget)
-            except (NoMatches, QueryError):
-                continue
-            if not getattr(control, "disabled", False):
-                control.focus()
-                return
+        return self._skills_controller._focus_library_skills_page_control(invoked)
 
-    def _refresh_library_skills_after_committed_mutation(
-        self,
-        *,
-        scope: SkillBrowseScope | None = None,
-    ) -> None:
-        """Invalidate exact totals after a durable Skill or trust change.
-
-        Retained rows may remain visible, but they are explicitly stale and
-        inert until the source-owned page refresh succeeds.
-        """
-        controller = self._library_skills_browse_controller
-        refresh_scope = scope or controller.mutation_refresh_scope
-        if controller.applied_result is not None:
-            controller.retain_stale_items(
-                controller.retained_items,
-                stale_copy="Skills changed; refresh the page before acting.",
-            )
-        else:
-            controller.invalidate(refresh_scope)
-        if (
-            self._library_selected_row_id == LIBRARY_ROW_BROWSE_SKILLS
-            and self._library_skills_view == "list"
-        ):
-            self._request_library_skills_browse(refresh_scope)
+    def _refresh_library_skills_after_committed_mutation(self, *, scope: SkillBrowseScope | None = None) -> None:
+        return self._skills_controller._refresh_library_skills_after_committed_mutation(scope=scope)
 
     def _refresh_library_skills_trust_posture(self) -> None:
         """Kick an off-thread read of the Skills trust service's posture.
@@ -18859,71 +18587,8 @@ class LibraryScreen(BaseAppScreen):
             exit_on_error=False,
         )
 
-    async def _load_library_skills_trust_posture(
-        self, posture_fn
-    ) -> LibraryEntryReconcileResult:
-        """Await the off-thread ``trust_posture()`` call and apply the result.
-
-        Args:
-            posture_fn: The trust service's bound ``trust_posture`` method
-                (captured by ``_refresh_library_skills_trust_posture`` so
-                this never re-reads ``local_skill_trust_service`` itself --
-                irrelevant here, but keeps this a pure "run this callable
-                off-thread" step).
-        """
-        route_key = self._library_entry_route_key()
-        generation = self._library_snapshot_state_generation
-        try:
-            posture = await asyncio.to_thread(posture_fn)
-        except Exception:
-            posture = ""
-        self._library_skills_trust_posture = posture if isinstance(posture, str) else ""
-        if (
-            route_key != self._library_entry_route_key()
-            or not self._library_entry_reconcile_is_current(generation, route_key)
-            or self._library_selected_row_id != LIBRARY_ROW_BROWSE_SKILLS
-            or self._library_skills_view != "list"
-        ):
-            return LibraryEntryReconcileResult.SUPERSEDED
-        identity = self._capture_library_entry_focus()
-        prior_callback: Callable[[], None] | None = None
-        pending = (generation, route_key)
-        if (
-            self._library_entry_reconcile_dirty
-            and self._library_entry_reconcile_pending == pending
-        ):
-            try:
-                canvas = self.query_one(
-                    "#library-skills-canvas", LibrarySkillsListCanvas
-                )
-            except (NoMatches, QueryError):
-                canvas = None
-            if canvas is not None:
-                prior_callback = canvas._post_recompose_callback
-
-        def finish_posture_sync() -> None:
-            if prior_callback is not None:
-                prior_callback()
-            if identity is not None:
-                self._restore_library_entry_focus(
-                    identity,
-                    generation=generation,
-                    route_key=route_key,
-                )
-
-        follow_up = (
-            finish_posture_sync
-            if prior_callback is not None or identity is not None
-            else None
-        )
-        if _sync_library_canvas(
-            self,
-            "skills",
-            then=follow_up,
-            allow_screen_fallback=False,
-        ):
-            return LibraryEntryReconcileResult.APPLIED
-        return LibraryEntryReconcileResult.FAILED
+    async def _load_library_skills_trust_posture(self, posture_fn) -> LibraryEntryReconcileResult:
+        return await self._skills_controller._load_library_skills_trust_posture(posture_fn)
 
     async def _refresh_library_media_detail(
         self,
@@ -25961,54 +25626,15 @@ class LibraryScreen(BaseAppScreen):
 
     @on(Button.Pressed, f"#{LIBRARY_SKILLS_PAGE_PREVIOUS_ID}")
     def handle_library_skills_page_previous(self, event: Button.Pressed) -> None:
-        """Request the preceding exact Skills page.
-
-        Args:
-            event: Button press event emitted by the Previous control.
-        """
-        event.stop()
-        controller = self._library_skills_browse_controller
-        applied = controller.applied_result
-        if applied is None or applied.page <= 1 or controller.freshness != "fresh":
-            return
-        self._request_library_skills_browse(
-            controller.scope_for_page(applied.page - 1),
-            focus_identity=LIBRARY_SKILLS_PAGE_PREVIOUS_ID,
-        )
+        return self._skills_controller.handle_library_skills_page_previous(event)
 
     @on(Button.Pressed, f"#{LIBRARY_SKILLS_PAGE_NEXT_ID}")
     def handle_library_skills_page_next(self, event: Button.Pressed) -> None:
-        """Request the following exact Skills page.
-
-        Args:
-            event: Button press event emitted by the Next control.
-        """
-        event.stop()
-        controller = self._library_skills_browse_controller
-        applied = controller.applied_result
-        if (
-            applied is None
-            or applied.page >= applied.total_pages
-            or controller.freshness != "fresh"
-        ):
-            return
-        self._request_library_skills_browse(
-            controller.scope_for_page(applied.page + 1),
-            focus_identity=LIBRARY_SKILLS_PAGE_NEXT_ID,
-        )
+        return self._skills_controller.handle_library_skills_page_next(event)
 
     @on(Button.Pressed, f"#{LIBRARY_SKILLS_RETRY_ID}")
     def handle_library_skills_retry(self, event: Button.Pressed) -> None:
-        """Retry the last requested Skills page or stale-page refresh.
-
-        Args:
-            event: Button press event emitted by the Retry control.
-        """
-        event.stop()
-        self._request_library_skills_browse(
-            self._library_skills_browse_controller.scope,
-            focus_identity=LIBRARY_SKILLS_RETRY_ID,
-        )
+        return self._skills_controller.handle_library_skills_retry(event)
 
     @on(Button.Pressed, "#library-skills-import")
     def handle_library_skills_import(self, event: Button.Pressed) -> None:
@@ -26241,28 +25867,12 @@ class LibraryScreen(BaseAppScreen):
             self._library_skill_import_coordinator.update_draft_path(event.value)
 
     @on(Input.Submitted, "#library-skills-import-path")
-    def handle_library_skills_import_path_submitted(
-        self, event: Input.Submitted
-    ) -> None:
-        """Run the import when Enter is pressed in the Import row's path field.
-
-        Args:
-            event: Input submission event emitted by the Import row's
-                path field.
-        """
-        event.stop()
-        self._start_library_skills_import()
+    def handle_library_skills_import_path_submitted(self, event: Input.Submitted) -> None:
+        return self._skills_controller.handle_library_skills_import_path_submitted(event)
 
     @on(Button.Pressed, "#library-skills-import-run")
     def handle_library_skills_import_run(self, event: Button.Pressed) -> None:
-        """Run the import when the Import row's "Import" action is pressed.
-
-        Args:
-            event: Button press event emitted by the Import row's
-                "Import" action.
-        """
-        event.stop()
-        self._start_library_skills_import()
+        return self._skills_controller.handle_library_skills_import_run(event)
 
     @on(Button.Pressed, "#library-skills-import-retry")
     def handle_library_skills_import_retry(self, event: Button.Pressed) -> None:
@@ -26441,198 +26051,28 @@ class LibraryScreen(BaseAppScreen):
         _sync_library_canvas(self, "skills")
 
     def _claim_library_skill_detail_generation(self) -> int:
-        """Start one Skill detail request and return its settlement fence."""
-        self._library_skill_detail_generation += 1
-        self._library_skill_detail_loading = True
-        self._library_skill_detail_error = ""
-        self._library_skill_detail_retryable = False
-        return self._library_skill_detail_generation
+        return self._skills_controller._claim_library_skill_detail_generation()
 
     def _invalidate_library_skill_detail_generation(self) -> None:
-        """Refuse every pending Skill detail settlement."""
-        self._library_skill_detail_generation += 1
-        self._library_skill_detail_loading = False
-        self._library_skill_detail_error = ""
-        self._library_skill_detail_retryable = False
+        return self._skills_controller._invalidate_library_skill_detail_generation()
 
-    def _library_skill_detail_request_is_current(
-        self, *, skill_name: str, generation: int
-    ) -> bool:
-        """Return whether one detail outcome still owns the Skills Work pane."""
-        return bool(
-            generation == self._library_skill_detail_generation
-            and skill_name == self._selected_skill_name
-            and self._library_skills_view == "editor"
-        )
+    def _library_skill_detail_request_is_current(self, *, skill_name: str, generation: int) -> bool:
+        return self._skills_controller._library_skill_detail_request_is_current(skill_name=skill_name, generation=generation)
 
     def _apply_library_skill_detail_failure(self, copy: str) -> None:
-        """Keep the selected Skill in place and expose a scoped retry."""
-        self._library_skill_detail_loading = False
-        self._library_skill_detail_error = copy
-        self._library_skill_detail_retryable = True
-        if self.is_mounted:
-            _sync_library_canvas(self, "skills")
+        return self._skills_controller._apply_library_skill_detail_failure(copy)
 
-    async def _refresh_library_skill_detail(
-        self,
-        skill_name: str,
-        *,
-        request_generation: int | None = None,
-    ) -> None:
-        """Fetch and store the full detail for a selected Library skill.
-
-        Mirrors ``_refresh_library_prompt_detail``: offloads the (possibly
-        blocking) ``get_skill`` service call via ``_run_library_service_call``
-        and recomposes once the fetched detail (or a cleared state) has
-        been stored.
-
-        Args:
-            skill_name: The Library skill name to fetch full detail for.
-        """
-        if request_generation is None:
-            request_generation = self._claim_library_skill_detail_generation()
-        service = getattr(self.app_instance, "skills_scope_service", None)
-        get_skill = getattr(service, "get_skill", None)
-        if not callable(get_skill):
-            if self._library_skill_detail_request_is_current(
-                skill_name=skill_name, generation=request_generation
-            ):
-                self._apply_library_skill_detail_failure(
-                    "Couldn’t load the selected Skill. The local service is unavailable."
-                )
-            return
-        failed = False
-        try:
-            detail = await self._run_library_service_call(
-                get_skill,
-                skill_name,
-                mode="local",
-                isolate_in_worker=True,
-            )
-        except Exception:
-            logger.opt(exception=True).warning(
-                f"Failed to load Library skill detail for {skill_name!r}."
-            )
-            detail = None
-            failed = True
-        # Discard out-of-order results: the same stale-race guard as
-        # ``_refresh_library_prompt_detail``.
-        if not self._library_skill_detail_request_is_current(
-            skill_name=skill_name, generation=request_generation
-        ):
-            return
-        if not isinstance(detail, Mapping):
-            self._apply_library_skill_detail_failure(
-                "Couldn’t load the selected Skill. Try again."
-                if failed
-                else "This Skill is no longer available. Refresh the list and try again."
-            )
-            return
-        self._library_skill_detail_loading = False
-        self._library_skill_detail_error = ""
-        self._library_skill_detail_retryable = False
-        self._apply_library_skill_detail(detail)
+    async def _refresh_library_skill_detail(self, skill_name: str, *, request_generation: int | None = None) -> None:
+        return await self._skills_controller._refresh_library_skill_detail(skill_name, request_generation=request_generation)
 
     def _apply_library_skill_detail(self, detail: Mapping[str, Any]) -> None:
-        """Store a freshly-fetched skill detail and (re)render the editor.
-
-        Shared by the initial open (``_refresh_library_skill_detail``) and
-        a successful Save (whose response mapping is already a full detail
-        -- see ``_save_library_skill``'s docstring for why no separate
-        "refresh snapshot" fetch is needed there).
-
-        Args:
-            detail: A skill detail mapping shaped like ``get_skill``'s (or
-                a save call's) response.
-        """
-        self._library_skill_detail = dict(detail)
-        self._library_skill_detail_loading = False
-        self._library_skill_detail_error = ""
-        self._library_skill_detail_retryable = False
-        self._library_skill_editor_state = build_skill_editor_state(
-            self._library_skill_detail
-        )
-        self._library_skill_original_name = self._library_skill_editor_state.name
-        self._library_skill_dirty = False
-        self._library_skill_status = ""
-        self._library_skill_conflict = False
-        self._library_skill_active_review = None
-        self._library_skill_script_grant = False
-        self._library_skill_tool_catalog = self._build_library_skill_tool_catalog()
-        self._library_skill_tool_filter = ""
-        self._library_skill_tool_captured = skill_allowed_tools_sequence(
-            self._library_skill_editor_state.allowed_tools_csv
-        )
-        self._library_skill_tool_picker_changed = False
-        self._library_skill_more_actions_open = False
-        self._library_skill_trust_details_open = False
-        self._library_skill_mutation_in_flight = False
-        self._library_skill_editor_armed = False
-        if self.is_mounted:
-            # Deterministic (task-15457 review I4b): arming is
-            # dirty-tracking, not cosmetics -- a follow-up lost to the
-            # canvas-pump race leaves the editor unarmed, so it rides the
-            # canvas's own post-recompose hook rather than the screen's.
-            _sync_library_canvas(self, "skills", then=self._arm_library_skill_editor)
-        # Task 7: not part of ``get_skill``'s response, so it needs its own
-        # off-thread fetch -- see ``_refresh_library_skill_script_grant``.
-        self._refresh_library_skill_script_grant()
+        return self._skills_controller._apply_library_skill_detail(detail)
 
     def _arm_library_skill_editor(self) -> None:
-        """Enable dirty-tracking once the skill editor's mount-time
-        ``Input.Changed``/``TextArea.Changed`` (fired for the non-empty
-        initial values) has already been delivered, so it is never mistaken
-        for a real edit.
-        """
-        self._library_skill_editor_armed = True
+        return self._skills_controller._arm_library_skill_editor()
 
     def _enter_library_skill_create_editor(self) -> None:
-        """Open the in-canvas skill editor on a blank, not-yet-saved record.
-
-        Entered via the Create rail's "New skill" row
-        (``LIBRARY_ROW_CREATE_SKILL``, whose ``target_id`` is ``"skills"``
-        -- the SAME canvas kind Browse > Skills targets), mirroring
-        ``_enter_library_prompt_create_editor``'s "New prompt" row.
-
-        ``_selected_skill_name`` stays ``""``: the sentinel
-        ``_save_library_skill`` already reads (``is_create = not name``)
-        to route its scope-service ``create_skill`` call instead of
-        ``update_skill``, and the sentinel ``compose_content`` reads
-        (``is_create=not self._selected_skill_name``) to keep the Name
-        Input editable (an existing skill's Name Input is disabled --
-        there is no rename primitive).
-
-        ``_library_skill_editor_state`` is built directly from an empty
-        mapping (``build_skill_editor_state({})``) rather than left
-        ``None``: ``compose_content``'s skills-editor branch gates on
-        ``_library_skill_editor_state is None`` to show a "Loading
-        skill…" placeholder while the async detail fetch
-        (``_refresh_library_skill_detail``) is in flight -- there is no
-        fetch for a brand-new record, so leaving it ``None`` would show
-        that placeholder forever.
-        """
-        self._selected_skill_name = ""
-        self._library_skills_view = "editor"
-        self._library_skill_reader_mode = "edit"
-        self._library_skill_detail = {}
-        self._invalidate_library_skill_detail_generation()
-        self._library_skill_editor_state = build_skill_editor_state({})
-        self._library_skill_original_name = ""
-        self._library_skill_dirty = False
-        self._library_skill_status = ""
-        self._library_skill_conflict = False
-        self._library_skill_active_review = None
-        self._library_skill_script_grant = False
-        self._library_skill_tool_catalog = self._build_library_skill_tool_catalog()
-        self._library_skill_tool_filter = ""
-        self._library_skill_tool_captured = ()
-        self._library_skill_tool_picker_changed = False
-        self._library_skill_more_actions_open = False
-        self._library_skill_trust_details_open = False
-        self._library_skill_mutation_in_flight = False
-        self._library_skill_confirming_delete = False
-        self._library_skill_scroll_pending = False
-        self._library_skill_editor_armed = False
+        return self._skills_controller._enter_library_skill_create_editor()
 
     def _reset_library_skill_editor_state(self) -> None:
         """Clear all in-canvas Library skill editor/save/trust state.
@@ -26664,372 +26104,66 @@ class LibraryScreen(BaseAppScreen):
         self._library_skill_editor_armed = False
 
     def _consume_library_skill_scroll_pending(self) -> bool:
-        """One-shot read of the create-save scroll-back flag (task-417).
-
-        Returns:
-            ``True`` exactly once per arm -- the first recompose after a
-            create-save scrolls the action row into view, later ones don't.
-        """
-        pending = self._library_skill_scroll_pending
-        self._library_skill_scroll_pending = False
-        return pending
+        return self._skills_controller._consume_library_skill_scroll_pending()
 
     def _library_skill_text_fields_match_state(self) -> bool:
-        """True when every live text field equals the editor state's value.
-
-        task-417: a recompose (e.g. the post-create snapshot refresh)
-        remounts the editor's Inputs/TextArea, whose spurious mount-time
-        ``Changed`` events can arrive AFTER any ``call_after_refresh``
-        re-arm -- the armed-flag dance alone cannot win that race. A mount
-        echo always carries the state's own values, so value equality is
-        the reliable spurious-vs-real discriminator.
-        """
-        state = self._library_skill_editor_state
-        if state is None:
-            return False
-        fields = self._read_library_skill_editor_fields()
-        if fields is None:
-            return False
-        (
-            raw_name,
-            raw_description,
-            raw_argument_hint,
-            raw_allowed_tools_csv,
-            raw_model,
-            raw_body,
-        ) = fields
-        return (
-            raw_name == (state.name or "")
-            and raw_description == (state.description or "")
-            and raw_argument_hint == (state.argument_hint or "")
-            and raw_allowed_tools_csv == (state.allowed_tools_csv or "")
-            and raw_model == (state.model or "")
-            and raw_body == (state.body or "")
-        )
+        return self._skills_controller._library_skill_text_fields_match_state()
 
     def _library_skill_on_disk_path(self) -> str:
-        """The selected skill's on-disk directory, for remediation copy.
-
-        Empty when the service or selection is unavailable (the copy
-        helper falls back to generic wording).
-        """
-        service = getattr(self.app_instance, "local_skills_service", None)
-        skills_dir = getattr(service, "skills_dir", None)
-        name = self._selected_skill_name
-        if skills_dir is None or not name:
-            return ""
-        try:
-            return str(Path(skills_dir) / name)
-        except Exception:
-            return ""
+        return self._skills_controller._library_skill_on_disk_path()
 
     def _mark_library_skill_dirty(self, *, force: bool = False) -> None:
-        """Record an in-progress skill edit.
-
-        Ignored until ``_library_skill_editor_armed`` is set (see that
-        flag's docstring), and -- unless ``force`` -- ignored when the live
-        text fields still equal the editor state (a mount-time ``Changed``
-        echo, task-417). The toggle/cycle buttons mutate the state BEFORE
-        marking, so they pass ``force=True``. Unlike the notes editor,
-        this never arms an autosave timer -- the skill editor is
-        explicit-Save-only.
-        """
-        if not self._library_skill_editor_armed:
-            return
-        if not force and self._library_skill_text_fields_match_state():
-            return
-        self._library_skill_dirty = True
-        # task-449: dirty-marking never recomposes, so the Discard button's
-        # initial disabled render is patched live here.
-        self._set_library_skill_discard_enabled(True)
-        # task-417: a lingering "Saved." is stale the moment new edits
-        # exist -- clear it alongside the dirty mark.
-        if self._library_skill_status:
-            self._update_library_skill_status_static("")
+        return self._skills_controller._mark_library_skill_dirty(force=force)
 
     def _read_library_skill_live_name(self) -> str:
-        """Read the Name Input's current (possibly unsaved) value.
-
-        Falls back to the editor state's own name when the Input isn't
-        mounted (e.g. the conflict banner, which doesn't render the field
-        Inputs).
-        """
-        try:
-            return self.query_one("#library-skill-name", Input).value
-        except (NoMatches, QueryError):
-            state = self._library_skill_editor_state
-            return state.name if state is not None else ""
+        return self._skills_controller._read_library_skill_live_name()
 
     def _update_library_skill_warnings_static(self, *, name: str | None = None) -> None:
-        """Targeted update of ``#library-skill-warnings``, no recompose.
-
-        Args:
-            name: The live Name field value to compute the shadow warning
-                against. Defaults to the current editor state's name (used
-                right after a trust action, where the Name field itself
-                hasn't changed).
-        """
-        state = self._library_skill_editor_state
-        if state is None:
-            return
-        if name is None:
-            name = state.name
-        try:
-            warnings_static = self.query_one("#library-skill-warnings", Static)
-        except (NoMatches, QueryError):
-            return
-        lines = skill_editor_warning_lines(
-            live_name=name,
-            trust_status=state.trust_status,
-            trust_blocked=state.trust_blocked,
-        )
-        warnings_static.update("\n".join(lines))
+        return self._skills_controller._update_library_skill_warnings_static(name=name)
 
     def _update_library_skill_status_static(self, text: str) -> None:
-        """Targeted update of ``#library-skill-save-status``, no recompose.
-
-        Args:
-            text: The status copy to show (``""`` clears it).
-        """
-        self._library_skill_status = text
-        try:
-            status_static = self.query_one("#library-skill-save-status", Static)
-        except (NoMatches, QueryError):
-            return
-        status_static.update(text)
+        return self._skills_controller._update_library_skill_status_static(text)
 
     def _render_library_skill_trust_panel(self) -> None:
-        """Targeted update of the trust panel's state/changed-files/buttons,
-        no recompose -- called after every trust action and after a
-        successful Save, so an in-progress (unsaved) edit elsewhere in the
-        editor is never discarded by a full rebuild.
-        """
-        state = self._library_skill_editor_state
-        if state is None:
-            return
-        try:
-            self.query_one("#library-skill-trust-state", Static).update(
-                skill_trust_state_line(state.trust_status, state.trust_changed_files)
-            )
-        except (NoMatches, QueryError):
-            pass
-        try:
-            self.query_one("#library-skill-trust-remediation", Static).update(
-                skill_trust_panel_remediation_copy(
-                    state.trust_status, self._library_skill_on_disk_path()
-                )
-            )
-        except (NoMatches, QueryError):
-            pass
-        try:
-            self.query_one("#library-skill-trust-review-identity", Static).update(
-                skill_review_identity_line(self._library_skill_active_review)
-            )
-        except (NoMatches, QueryError):
-            pass
-        try:
-            self.query_one("#library-skill-trust-review-files", Static).update(
-                ", ".join(
-                    str(item)
-                    for item in (
-                        (self._library_skill_active_review or {}).get("changed_files")
-                        or []
-                    )
-                )
-            )
-        except (NoMatches, QueryError):
-            pass
-        try:
-            self.query_one("#library-skill-trust-review-content", Static).update(
-                skill_trust_review_preview(self._library_skill_active_review)
-            )
-        except (NoMatches, QueryError):
-            pass
-        try:
-            unlock_button = self.query_one("#library-skill-trust-unlock", Button)
-            unlock_button.disabled = not skill_trust_unlock_enabled(state.trust_status)
-            # F-018: reason/action tooltips flip in place with `disabled`.
-            unlock_button.tooltip = skill_trust_unlock_tooltip(state.trust_status)
-        except (NoMatches, QueryError):
-            pass
-        try:
-            review_button = self.query_one("#library-skill-trust-review", Button)
-            review_button.disabled = not skill_trust_review_enabled(
-                state.trust_status, state.trust_blocked
-            )
-            review_button.tooltip = skill_trust_review_tooltip(
-                state.trust_status, state.trust_blocked
-            )
-        except (NoMatches, QueryError):
-            pass
-        try:
-            approve_button = self.query_one("#library-skill-trust-approve", Button)
-            approve_button.disabled = self._library_skill_active_review is None
-            approve_button.tooltip = skill_trust_approve_tooltip(
-                self._library_skill_active_review is not None
-            )
-        except (NoMatches, QueryError):
-            pass
-        # Task 7 (skills-script-execution): reads the CACHED grant
-        # (``_library_skill_script_grant``), never the trust service
-        # directly -- ``script_execution_granted`` re-scans the skill's
-        # on-disk directory to verify its fingerprint, which is blocking
-        # file I/O this method (called from synchronous event handlers) must
-        # not perform. ``_refresh_library_skill_script_grant`` is what keeps
-        # the cache current, off-thread.
-        try:
-            self.query_one("#library-skill-script-grant", Static).update(
-                skill_script_grant_line(self._library_skill_script_grant)
-            )
-            self.query_one(
-                "#library-skill-script-grant-revoke", Button
-            ).disabled = not self._library_skill_script_grant
-        except (NoMatches, QueryError):
-            pass
+        return self._skills_controller._render_library_skill_trust_panel()
 
     @on(Input.Changed, "#library-skill-name")
     def handle_library_skill_name_changed(self, event: Input.Changed) -> None:
-        """Mark the open skill dirty on a Name edit, and live-refresh the
-        shadow-name warning (unconditionally, not gated by "armed" -- the
-        warning is a plain live read, not a dirty-tracking concern).
-
-        Args:
-            event: Input change event emitted by the editor's Name field.
-        """
-        self._mark_library_skill_dirty()
-        self._update_library_skill_warnings_static(name=event.value)
+        return self._skills_controller.handle_library_skill_name_changed(event)
 
     @on(Input.Changed, "#library-skill-argument-hint")
     @on(Input.Changed, "#library-skill-allowed-tools")
     def handle_library_skill_input_changed(self, event: Input.Changed) -> None:
-        """Mark the open skill dirty on a field edit.
-
-        Args:
-            event: Input change event emitted by one of the editor's
-                single-line fields.
-        """
-        self._mark_library_skill_dirty()
+        return self._skills_controller.handle_library_skill_input_changed(event)
 
     @on(Input.Changed, "#library-skill-description")
     def handle_library_skill_description_changed(self, event: Input.Changed) -> None:
-        """Mark dirty on a Description edit and sync the derived-hint (review
-        finding).
-
-        The "No description set" hint (rendered only when the skill's
-        description was auto-derived from its body) is compose-time-only, so
-        without this it would linger beneath a now-populated field. It is
-        meaningful only while the field is still empty.
-
-        Args:
-            event: Input change event emitted by the Description field.
-        """
-        self._mark_library_skill_dirty()
-        self._sync_library_skill_description_hint(event.value)
+        return self._skills_controller.handle_library_skill_description_changed(event)
 
     def _sync_library_skill_description_hint(self, live_value: str) -> None:
-        """Show the derived-description hint only while the field is empty."""
-        state = self._library_skill_editor_state
-        should_show = bool(
-            state is not None and state.description_derived and not live_value.strip()
-        )
-        for hint in self.query("#library-skill-description-hint"):
-            if isinstance(hint, Static):
-                hint.display = should_show
+        return self._skills_controller._sync_library_skill_description_hint(live_value)
 
     @on(TextArea.Changed, "#library-skill-body")
     def handle_library_skill_body_changed(self, event: TextArea.Changed) -> None:
-        """Mark the open skill dirty on a Body edit.
-
-        Args:
-            event: Text change event emitted by the editor's Body TextArea.
-        """
-        self._mark_library_skill_dirty()
+        return self._skills_controller.handle_library_skill_body_changed(event)
 
     def _update_library_skill_toggle_buttons(self) -> None:
-        """Targeted label update for the user-invocable/disable-model/context
-        toggle Buttons, no recompose."""
-        state = self._library_skill_editor_state
-        if state is None:
-            return
-        try:
-            self.query_one(
-                "#library-skill-user-invocable", Button
-            ).label = skill_user_invocable_label(state.user_invocable)
-        except (NoMatches, QueryError):
-            pass
-        try:
-            self.query_one("#library-skill-invocation-copy", Static).update(
-                skill_invocation_copy(
-                    state.user_invocable, state.disable_model_invocation
-                )
-            )
-            self.query_one("#library-skill-argument-fields").display = bool(
-                self._library_skill_editor_mode == "advanced" or state.user_invocable
-            )
-        except (NoMatches, QueryError):
-            pass
+        return self._skills_controller._update_library_skill_toggle_buttons()
 
     @on(Button.Pressed, "#library-skill-editor-mode")
     async def handle_library_skill_editor_mode(self, event: Button.Pressed) -> None:
-        """Switch mounted Skill presentations and remember the display choice."""
-        event.stop()
-        self._snapshot_library_skill_live_fields()
-        requested = (
-            "basic" if self._library_skill_editor_mode == "advanced" else "advanced"
-        )
-        try:
-            canvas = self.query_one("#library-skill-work-pane", LibrarySkillWorkPane)
-        except NoMatches:
-            return
-        await canvas.set_editor_mode(requested)
-        self._library_skill_editor_mode = requested
-        library_config = self.app_instance.app_config.setdefault("library", {})
-        if isinstance(library_config, dict):
-            library_config["skill_editor_mode"] = requested
-        self.run_worker(
-            self._persist_library_skill_editor_mode(requested),
-            group="library_skill_editor_mode",
-            exclusive=True,
-        )
+        return await self._skills_controller.handle_library_skill_editor_mode(event)
 
     @on(Button.Pressed, "#library-skill-mode-overview")
     @on(Button.Pressed, "#library-skill-mode-edit")
     @on(Button.Pressed, "#library-skill-mode-trust")
     @on(Button.Pressed, "#library-skill-mode-files")
     def handle_library_skill_reader_mode(self, event: Button.Pressed) -> None:
-        """Switch the selected Skill's explicit work-pane projection."""
-        event.stop()
-        requested = coerce_skill_reader_mode(event.button.id.rsplit("-", 1)[-1])
-        if not self._selected_skill_name and requested != "edit":
-            return
-        if self._library_skill_reader_mode == "edit":
-            self._snapshot_library_skill_live_fields()
-        if requested == self._library_skill_reader_mode:
-            return
-        self._library_skill_reader_mode = requested
-        _sync_library_canvas(self, "skills")
+        return self._skills_controller.handle_library_skill_reader_mode(event)
 
     @on(Button.Pressed, "#library-skill-detail-retry")
     def handle_library_skill_detail_retry(self, event: Button.Pressed) -> None:
-        """Retry only the still-selected Skill detail with a fresh fence."""
-        event.stop()
-        name = self._selected_skill_name
-        if (
-            not name
-            or self._library_skill_mutation_in_flight
-            or not self._library_skill_detail_retryable
-        ):
-            return
-        generation = self._claim_library_skill_detail_generation()
-        self.run_worker(
-            self._refresh_library_skill_detail(
-                name,
-                request_generation=generation,
-            ),
-            exclusive=True,
-            group="library_skill_detail",
-        )
-        _sync_library_canvas(self, "skills")
+        return self._skills_controller.handle_library_skill_detail_retry(event)
 
     async def _persist_library_skill_editor_mode(self, mode: str) -> None:
         """Persist one admitted Skill display mode without blocking Textual."""
@@ -27050,144 +26184,30 @@ class LibraryScreen(BaseAppScreen):
 
     @on(Input.Changed, "#library-skill-tool-filter")
     def handle_library_skill_tool_filter(self, event: Input.Changed) -> None:
-        """Filter picker rows without changing the Skill's allowlist content."""
-        self._library_skill_tool_filter = event.value
-        try:
-            self.query_one(
-                "#library-skill-work-pane", LibrarySkillWorkPane
-            ).set_tool_filter(event.value)
-        except (NoMatches, QueryError):
-            return
+        return self._skills_controller.handle_library_skill_tool_filter(event)
 
     @on(SelectionList.SelectedChanged, "#library-skill-tool-picker")
-    def handle_library_skill_tool_selection(
-        self, event: SelectionList.SelectedChanged
-    ) -> None:
-        """Apply only a user-driven tool restriction edit to the canonical draft."""
-        try:
-            canvas = self.query_one("#library-skill-work-pane", LibrarySkillWorkPane)
-        except (NoMatches, QueryError):
-            return
-        if canvas.rebuilding_tool_picker or not self._library_skill_editor_armed:
-            return
-        state = self._library_skill_editor_state
-        if state is None:
-            return
-        visible = {selection.value for selection in event.selection_list.options}
-        selected = set(skill_allowed_tools_sequence(state.allowed_tools_csv))
-        selected.difference_update(visible)
-        selected.update(str(value) for value in event.selection_list.selected)
-        current_selected = set(skill_allowed_tools_sequence(state.allowed_tools_csv))
-        if selected == current_selected:
-            return
-        reconciled = reconcile_skill_allowed_tools(
-            self._library_skill_tool_captured,
-            selected=tuple(selected),
-            catalog_order=self._library_skill_tool_catalog,
-            picker_changed=True,
-        )
-        self._library_skill_tool_picker_changed = True
-        self._library_skill_editor_state = dataclasses.replace(
-            state, allowed_tools_csv=", ".join(reconciled)
-        )
-        self._mark_library_skill_dirty(force=True)
-        try:
-            self.query_one(
-                "#library-skill-disable-model", Button
-            ).label = skill_disable_model_label(state.disable_model_invocation)
-        except (NoMatches, QueryError):
-            pass
-        try:
-            self.query_one(
-                "#library-skill-context", Button
-            ).label = skill_context_toggle_label(state.context)
-        except (NoMatches, QueryError):
-            pass
+    def handle_library_skill_tool_selection(self, event: SelectionList.SelectedChanged) -> None:
+        return self._skills_controller.handle_library_skill_tool_selection(event)
 
     @on(Button.Pressed, "#library-skill-user-invocable")
     def handle_library_skill_user_invocable_toggle(self, event: Button.Pressed) -> None:
-        """Toggle the open skill's ``user_invocable`` flag.
-
-        Args:
-            event: Button press event emitted by the user-invocable toggle.
-        """
-        event.stop()
-        state = self._library_skill_editor_state
-        if state is None:
-            return
-        self._library_skill_editor_state = dataclasses.replace(
-            state, user_invocable=not state.user_invocable
-        )
-        self._mark_library_skill_dirty(force=True)
-        self._update_library_skill_toggle_buttons()
+        return self._skills_controller.handle_library_skill_user_invocable_toggle(event)
 
     @on(Button.Pressed, "#library-skill-disable-model")
     def handle_library_skill_disable_model_toggle(self, event: Button.Pressed) -> None:
-        """Toggle the open skill's ``disable_model_invocation`` flag.
-
-        Args:
-            event: Button press event emitted by the disable-model toggle.
-        """
-        event.stop()
-        state = self._library_skill_editor_state
-        if state is None:
-            return
-        self._library_skill_editor_state = dataclasses.replace(
-            state, disable_model_invocation=not state.disable_model_invocation
-        )
-        self._mark_library_skill_dirty(force=True)
-        self._update_library_skill_toggle_buttons()
+        return self._skills_controller.handle_library_skill_disable_model_toggle(event)
 
     @on(Button.Pressed, "#library-skill-context")
     def handle_library_skill_context_toggle(self, event: Button.Pressed) -> None:
-        """Cycle the open skill's ``context`` field between ``inline``/``fork``.
+        return self._skills_controller.handle_library_skill_context_toggle(event)
 
-        Args:
-            event: Button press event emitted by the context cycler.
-        """
-        event.stop()
-        state = self._library_skill_editor_state
-        if state is None:
-            return
-        self._library_skill_editor_state = dataclasses.replace(
-            state, context=next_skill_context(state.context)
-        )
-        self._mark_library_skill_dirty(force=True)
-        self._update_library_skill_toggle_buttons()
-
-    def _read_library_skill_editor_fields(
-        self,
-    ) -> tuple[str, str, str, str, str, str] | None:
-        """Read the skill editor's current (possibly unsaved) field values.
-
-        Returns:
-            ``(name, description, argument_hint, allowed_tools_csv, model,
-            body)`` read from the live widgets, or ``None`` if the editor
-            isn't mounted.
-        """
-        try:
-            name = self.query_one("#library-skill-name", Input).value
-            description = self.query_one("#library-skill-description", Input).value
-            argument_hint = self.query_one("#library-skill-argument-hint", Input).value
-            body = self.query_one("#library-skill-body", TextArea).text
-        except (NoMatches, QueryError):
-            return None
-        state = self._library_skill_editor_state
-        if state is None:
-            return None
-        allowed_tools_csv = state.allowed_tools_csv
-        model = state.model or ""
-        return name, description, argument_hint, allowed_tools_csv, model, body
+    def _read_library_skill_editor_fields(self) -> tuple[str, str, str, str, str, str] | None:
+        return self._skills_controller._read_library_skill_editor_fields()
 
     @on(Button.Pressed, "#library-skill-save")
     def handle_library_skill_save(self, event: Button.Pressed) -> None:
-        """Explicitly save the open skill (there is no autosave).
-
-        Args:
-            event: Button press event emitted by the editor's "Save" action.
-        """
-        event.stop()
-        self._begin_library_skill_save()
+        return self._skills_controller.handle_library_skill_save(event)
 
     def _begin_library_skill_save(self) -> bool:
         """Admit one save and expose its shared lifecycle interlock."""
@@ -27204,329 +26224,32 @@ class LibraryScreen(BaseAppScreen):
         return True
 
     async def _run_library_skill_save(self) -> None:
-        """Hold the shared editor interlock for one durable Skill save."""
-        try:
-            await self._save_library_skill()
-        finally:
-            self._library_skill_mutation_in_flight = False
-            if self.is_mounted:
-                self._sync_library_skill_lifecycle_actions()
+        return await self._skills_controller._run_library_skill_save()
 
     async def _save_library_skill(self) -> None:
-        """Save the open Library skill's current editor text.
+        return await self._skills_controller._save_library_skill()
 
-        Unlike the prompts editor (whose ``update_prompt_by_id`` has no
-        caller-supplied expected-version parameter, forcing a manual
-        pre-read staleness check), ``LocalSkillsService.update_skill``
-        accepts ``expected_version`` directly and raises
-        ``local_skill_version_conflict:...`` itself on a real mismatch --
-        so this never needs its own pre-read; ``classify_skill_save_error``
-        classifies whatever the real write call raises/returns.
-
-        The create/update response mapping is already a full skill detail
-        (``LocalSkillsService._response_for_record``'s shape, same as
-        ``get_skill``'s), so a successful save's "refresh snapshot" is just
-        rebuilding the editor state from THIS call's own result -- no
-        second service round-trip needed. This is also how the
-        save-marks-needs-review re-quarantine becomes visible without any
-        special-casing: the write never passes ``trust_approved=True``, so
-        a currently-trusted skill's post-save ``trust_status`` in the
-        response is already ``quarantined_modified``.
-        """
-        if self._library_skills_view != "editor":
-            return
-        name = self._selected_skill_name
-        is_create = not name
-        base_state = self._library_skill_editor_state
-        if base_state is None:
-            return
-        fields = self._read_library_skill_editor_fields()
-        if fields is None:
-            return
-        (
-            raw_name,
-            raw_description,
-            raw_argument_hint,
-            raw_allowed_tools_csv,
-            raw_model,
-            raw_body,
-        ) = fields
-
-        live_name = self._sanitize_media_field(raw_name, max_length=64)
-        description = self._sanitize_note_content(
-            raw_description, max_length=LIBRARY_SKILL_TEXT_MAX_CHARS
-        )
-        argument_hint = self._sanitize_media_field(raw_argument_hint, max_length=500)
-        allowed_tools_csv = self._sanitize_note_content(
-            raw_allowed_tools_csv, max_length=LIBRARY_SKILL_TEXT_MAX_CHARS
-        )
-        model = self._sanitize_media_field(raw_model, max_length=128)
-        body = self._sanitize_note_content(
-            raw_body, max_length=LIBRARY_SKILL_TEXT_MAX_CHARS
-        )
-
-        if is_create:
-            editor_name = live_name or base_state.name
-        else:
-            # Renaming an existing skill isn't supported -- the service has
-            # no rename primitive, and ``update_skill`` writes under the
-            # ORIGINAL directory name regardless of what the frontmatter's
-            # ``name`` field says. The Name Input is disabled for existing
-            # skills (see ``LibrarySkillsListCanvas._compose_editor``), but
-            # this pins the persisted name defensively too: even if the
-            # live value somehow diverged, the frontmatter written to disk
-            # never does, so a save can never get marked
-            # ``validation_status: "invalid"`` (name != parent directory
-            # name) the way it silently did before this fix.
-            editor_name = base_state.name
-
-        write_state = dataclasses.replace(
-            base_state,
-            name=editor_name,
-            description=description,
-            argument_hint=argument_hint or None,
-            allowed_tools_csv=allowed_tools_csv,
-            model=model or None,
-        )
-        content = compose_skill_markdown(write_state, body=body)
-
-        service = getattr(self.app_instance, "skills_scope_service", None)
-        create_skill = getattr(service, "create_skill", None)
-        update_skill = getattr(service, "update_skill", None)
-
-        result: Any = None
-        exc: Exception | None = None
-        if is_create:
-            if not callable(create_skill):
-                return
-            try:
-                result = await self._run_library_service_call(
-                    create_skill,
-                    mode="local",
-                    name=write_state.name,
-                    content=content,
-                    isolate_in_worker=True,
-                )
-            except Exception as caught:
-                exc = caught
-        else:
-            if not callable(update_skill):
-                return
-            try:
-                result = await self._run_library_service_call(
-                    update_skill,
-                    name,
-                    mode="local",
-                    content=content,
-                    expected_version=base_state.version,
-                    isolate_in_worker=True,
-                )
-            except Exception as caught:
-                exc = caught
-
-        # Discard out-of-order results, same stale-race guard as
-        # ``_refresh_library_skill_detail``/``_save_library_prompt``'s
-        # equivalent ``prompt_id != self._selected_prompt_id`` check --
-        # applied uniformly for creates too (``name`` was already ``""``
-        # at capture time when ``is_create``, so this still lets a
-        # still-in-flight create through as long as nothing else got
-        # selected meanwhile, but bails if a DIFFERENT skill's editor
-        # opened while this create was in flight).
-        if name != self._selected_skill_name or self._library_skills_view != "editor":
-            return
-
-        if exc is not None:
-            logger.opt(exception=True).warning(
-                f"Library skill save failed for {name!r}."
-            )
-            outcome = classify_skill_save_error(None, str(exc), exc)
-        else:
-            outcome = classify_skill_save_error(result, "", None)
-
-        if outcome == "version-conflict":
-            self._enter_library_skill_conflict()
-            return
-        if outcome != "ok":
-            self._update_library_skill_status_static(
-                LIBRARY_SKILL_SAVE_STATUS_COPY.get(
-                    outcome, LIBRARY_SKILL_SAVE_STATUS_COPY["error"]
-                )
-            )
-            return
-
-        self._apply_library_skill_save_success(result, is_create=is_create)
-
-    def _apply_library_skill_save_success(
-        self, result: Any, *, is_create: bool
-    ) -> None:
-        """Apply a successful save's response: rebuild state, clear dirty,
-        show "Saved.", and refresh the trust panel + warnings in place.
-
-        Args:
-            result: The create/update call's response mapping (a full
-                skill detail).
-            is_create: Whether this save created a brand-new skill (adopts
-                the new name as ``_selected_skill_name`` and kicks a
-                snapshot refresh so the rail badge/list pick up the new
-                row).
-        """
-        if not isinstance(result, Mapping):
-            self._update_library_skill_status_static(
-                LIBRARY_SKILL_SAVE_STATUS_COPY["error"]
-            )
-            return
-        # Deliberately NOT ``_apply_library_skill_detail`` (which recomposes
-        # + re-arms): recomposing here would remount fresh Input/TextArea
-        # widgets while the editor is still armed, and Textual's spurious
-        # mount-time ``Changed`` event for a non-empty initial value would
-        # immediately re-mark the just-saved skill dirty -- same discipline
-        # ``_save_library_prompt``'s success tail documents.
-        self._library_skill_detail = dict(result)
-        self._library_skill_editor_state = build_skill_editor_state(
-            self._library_skill_detail
-        )
-        self._library_skill_tool_captured = skill_allowed_tools_sequence(
-            self._library_skill_editor_state.allowed_tools_csv
-        )
-        self._library_skill_tool_picker_changed = False
-        self._library_skill_original_name = self._library_skill_editor_state.name
-        self._library_skill_dirty = False
-        # A content save changes the exact files/fingerprint that any prior
-        # trust receipt identified. Never leave that receipt approvable.
-        self._library_skill_active_review = None
-        self._refresh_library_skills_after_committed_mutation()
-        # task-449: this success tail deliberately never recomposes, so the
-        # Discard button is re-disabled in place alongside the dirty clear.
-        self._set_library_skill_discard_enabled(False)
-        if is_create:
-            self._selected_skill_name = self._library_skill_editor_state.name
-            # task-417: the snapshot refresh below lands a recompose that
-            # would reset the canvas scroll to the top, away from the Save
-            # button the user just pressed -- arm the one-shot scroll-back.
-            self._library_skill_scroll_pending = True
-            # A brand-new skill changes the list's membership/count, so the
-            # Skills rail badge and list must pick up the new row now --
-            # fire-and-forget, mirrors ``_save_library_prompt``'s equivalent
-            # post-create refresh.
-            self._refresh_local_source_snapshot()
-        self._update_library_skill_status_static(
-            (
-                "Saved. Review trust before using this Skill with the agent."
-                if is_create
-                else LIBRARY_SKILL_SAVE_STATUS_COPY["ok"]
-            )
-        )
-        self._update_library_skill_warnings_static(
-            name=self._read_library_skill_live_name()
-        )
-        self._render_library_skill_trust_panel()
-        if is_create and self.is_mounted:
-            # The first durable save changes the editor's structural truth:
-            # it gains a real trust panel and saved-clean lifecycle. Replace
-            # the canvas once at that commit point, never for display-mode or
-            # ordinary lifecycle changes.
-            self._library_skill_editor_armed = False
-            _sync_library_canvas(
-                self,
-                "skills",
-                then=self._arm_library_skill_editor,
-            )
-        # Task 7: saved content changes the skill's fingerprint, which
-        # invalidates any prior standing script grant -- re-check off-thread
-        # rather than let a just-invalidated grant keep showing as active.
-        self._refresh_library_skill_script_grant()
+    def _apply_library_skill_save_success(self, result: Any, *, is_create: bool) -> None:
+        return self._skills_controller._apply_library_skill_save_success(result, is_create=is_create)
 
     def _enter_library_skill_conflict(self) -> None:
-        """Recompose into the save-conflict banner (Reload only -- see the
-        brief's narrower scope vs. the prompts editor's Overwrite+Reload;
-        ``update_skill``'s own ``expected_version`` guard is what raised
-        this, so nothing here needs to preserve/replay the user's kept
-        text the way the prompts conflict path does).
-        """
-        self._library_skill_conflict = True
-        self._library_skill_status = ""
-        if self.is_mounted:
-            self._sync_library_skill_lifecycle_actions()
+        return self._skills_controller._enter_library_skill_conflict()
 
     @on(Button.Pressed, "#library-skill-conflict-reload")
     def handle_library_skill_conflict_reload(self, event: Button.Pressed) -> None:
-        """Discard the conflicting edit and refetch the skill's fresh detail.
-
-        Args:
-            event: Button press event emitted by the conflict banner's
-                "Reload" action.
-        """
-        event.stop()
-        name = self._selected_skill_name
-        if not name:
-            return
-        self._library_skill_conflict = False
-        self.run_worker(
-            self._refresh_library_skill_detail(name),
-            exclusive=True,
-            group="library_skill_detail",
-        )
+        return self._skills_controller.handle_library_skill_conflict_reload(event)
 
     async def _flush_library_skill_save(self) -> bool:
-        """Veto leaving the skill editor while an edit is unsaved.
-
-        Mirrors ``_flush_library_prompt_save`` exactly: the skill editor is
-        explicit-Save-only, so this simply reports whether it is safe to
-        proceed -- ``False`` whenever ``_library_skill_dirty`` is set.
-
-        Returns:
-            ``True`` when there is nothing unsaved (safe to proceed);
-            ``False`` when a dirty edit must be resolved first.
-        """
-        return not self._library_skill_dirty
+        return await self._skills_controller._flush_library_skill_save()
 
     def _notify_skill_dirty_veto(self) -> None:
-        """Tell the user WHY a skill-editor exit was just vetoed (task-449).
-
-        Every ``_flush_library_skill_save`` veto site calls this so a
-        blocked Back / skill-row / rail-row click never reads as a dead
-        button. Warning severity matches the trust-action toasts.
-        """
-        notify = getattr(self.app_instance, "notify", None)
-        if callable(notify):
-            notify(LIBRARY_SKILL_DIRTY_VETO_COPY, severity="warning")
+        return self._skills_controller._notify_skill_dirty_veto()
 
     def _set_library_skill_discard_enabled(self, enabled: bool) -> None:
-        """Patch the Discard button's disabled state in place (task-449).
-
-        Dirty-marking and the save-success tail both deliberately avoid a
-        recompose (see ``_apply_library_skill_save_success``), so the
-        Discard button's initial ``disabled=not dirty`` render must be
-        kept current by hand at those two transitions.
-        """
-        self._sync_library_skill_lifecycle_actions()
-        for button in self.query("#library-skill-discard"):
-            if isinstance(button, Button):
-                button.disabled = not enabled
-                # F-018: the reason/action tooltip flips in place with
-                # `disabled` (this patcher exists precisely to avoid a
-                # recompose, so the compose-time tooltip would go stale).
-                button.tooltip = (
-                    SKILL_DISCARD_TOOLTIP_DIRTY
-                    if enabled
-                    else SKILL_DISCARD_TOOLTIP_CLEAN
-                )
+        return self._skills_controller._set_library_skill_discard_enabled(enabled)
 
     def _sync_library_skill_lifecycle_actions(self) -> bool:
-        """Patch the mounted Skill action strip from screen-owned state."""
-        try:
-            canvas = self.query_one("#library-skill-work-pane", LibrarySkillWorkPane)
-            canvas.sync_lifecycle_actions(
-                dirty=self._library_skill_dirty,
-                conflict=self._library_skill_conflict,
-                confirming_delete=self._library_skill_confirming_delete,
-                mutation_in_flight=self._library_skill_mutation_in_flight,
-                more_actions_open=self._library_skill_more_actions_open,
-                is_create=not self._selected_skill_name,
-            )
-        except (NoMatches, QueryError):
-            return False
-        return True
+        return self._skills_controller._sync_library_skill_lifecycle_actions()
 
     def _library_skill_editor_active(self) -> bool:
         """True while the in-canvas skill editor is the live view (task-424)."""
@@ -28009,11 +26732,7 @@ class LibraryScreen(BaseAppScreen):
         return True
 
     def _focus_library_skill_name(self) -> None:
-        """Focus the create editor's Name field (task-424)."""
-        try:
-            self.query_one("#library-skill-name", Input).focus()
-        except (NoMatches, QueryError):
-            pass
+        return self._skills_controller._focus_library_skill_name()
 
     def _focus_library_ingest_path(self) -> None:
         """Focus the Ingest canvas's path field (task-3302 AC#1, MI-03).
@@ -28117,72 +26836,23 @@ class LibraryScreen(BaseAppScreen):
 
     @on(Button.Pressed, "#library-skill-back")
     async def handle_library_skill_back(self, event: Button.Pressed) -> None:
-        """Return the Library skills canvas from the editor to its list view.
-
-        Vetoed while dirty (see ``_flush_library_skill_save``) so Back
-        never silently discards an unsaved edit.
-
-        Args:
-            event: Button press event emitted by the "‹ Back to list" action.
-        """
-        event.stop()
-        await self._exit_library_skill_editor_guarded()
+        return await self._skills_controller.handle_library_skill_back(event)
 
     @on(Button.Pressed, "#library-skill-cancel")
     def handle_library_skill_cancel(self, event: Button.Pressed) -> None:
-        """Cancel a never-saved Skill draft and return to the list."""
-        event.stop()
-        if self._selected_skill_name or self._library_skill_mutation_in_flight:
-            return
-        self._reset_library_skill_editor_state()
-        self._refresh_local_source_snapshot()
-        _sync_library_canvas(self, "skills")
+        return self._skills_controller.handle_library_skill_cancel(event)
 
     @on(Button.Pressed, "#library-skill-more-actions")
     def handle_library_skill_more_actions(self, event: Button.Pressed) -> None:
-        """Reveal saved-clean secondary actions without changing the draft."""
-        event.stop()
-        if self._library_skill_dirty or self._library_skill_mutation_in_flight:
-            return
-        self._snapshot_library_skill_live_fields()
-        self._library_skill_more_actions_open = (
-            not self._library_skill_more_actions_open
-        )
-        self._sync_library_skill_lifecycle_actions()
-        try:
-            self.query_one("#library-skill-more-actions", Button).focus()
-        except (NoMatches, QueryError):
-            pass
+        return self._skills_controller.handle_library_skill_more_actions(event)
 
     @on(Button.Pressed, "#library-skill-trust-view-details")
     def handle_library_skill_trust_view_details(self, event: Button.Pressed) -> None:
-        """Open the healthy trust detail region without changing editor mode."""
-        event.stop()
-        self._snapshot_library_skill_live_fields()
-        self._library_skill_trust_details_open = True
-        _sync_library_canvas(self, "skills")
+        return self._skills_controller.handle_library_skill_trust_view_details(event)
 
     @on(Button.Pressed, "#library-skill-discard")
     def handle_library_skill_discard(self, event: Button.Pressed) -> None:
-        """Leave the skill editor WITHOUT saving the dirty edit (task-449).
-
-        The explicit counterpart to the dirty vetoes: Back and every row
-        switch refuse to move while ``_library_skill_dirty`` is set, so
-        this button is the one deliberate way out that drops the edit.
-        Same exit tail as a clean Back (reset + snapshot + recompose); the
-        button renders disabled until the editor is actually dirty, so a
-        stray click on a clean editor can't discard anything.
-
-        Args:
-            event: Button press event emitted by the "Discard changes"
-                action.
-        """
-        event.stop()
-        if not self._library_skill_dirty:
-            return
-        self._reset_library_skill_editor_state()
-        self._refresh_local_source_snapshot()
-        _sync_library_canvas(self, "skills")
+        return self._skills_controller.handle_library_skill_discard(event)
 
     @on(Button.Pressed, "#library-skill-delete")
     def handle_library_skill_delete(self, event: Button.Pressed) -> None:
@@ -28210,42 +26880,7 @@ class LibraryScreen(BaseAppScreen):
                 pass
 
     def _snapshot_library_skill_live_fields(self) -> None:
-        """Fold the editor's live (possibly unsaved) field values back into
-        ``_library_skill_editor_state`` so a state-driven recompose renders
-        exactly what the user had typed (task-415's confirm step).
-        """
-        state = self._library_skill_editor_state
-        if state is None:
-            return
-        fields = self._read_library_skill_editor_fields()
-        if fields is None:
-            return
-        (
-            raw_name,
-            raw_description,
-            raw_argument_hint,
-            raw_allowed_tools_csv,
-            raw_model,
-            raw_body,
-        ) = fields
-        self._library_skill_editor_state = dataclasses.replace(
-            state,
-            # Rename is unsupported; existing skills keep their name (the
-            # Name Input is disabled there anyway), create mode keeps the
-            # typed name.
-            name=raw_name if not self._selected_skill_name else state.name,
-            description=raw_description,
-            # A typed description is no longer "derived from the body", so a
-            # state-driven recompose must not re-show the "No description
-            # set" hint next to the populated field (review finding). Stays
-            # derived only while the field is still empty.
-            description_derived=state.description_derived
-            and not raw_description.strip(),
-            argument_hint=raw_argument_hint,
-            allowed_tools_csv=raw_allowed_tools_csv,
-            model=raw_model,
-            body=raw_body,
-        )
+        return self._skills_controller._snapshot_library_skill_live_fields()
 
     @on(Button.Pressed, "#library-skill-delete-confirm")
     def handle_library_skill_delete_confirm(self, event: Button.Pressed) -> None:
@@ -28273,23 +26908,8 @@ class LibraryScreen(BaseAppScreen):
             group="library_skill_delete",
         )
 
-    async def _run_library_skill_delete(
-        self, skill_name: str, request_generation: int
-    ) -> None:
-        """Hold the shared editor interlock for one durable Skill delete."""
-        try:
-            await self._delete_library_skill(
-                skill_name,
-                request_generation=request_generation,
-            )
-        finally:
-            if self._library_skill_detail_request_is_current(
-                skill_name=skill_name,
-                generation=request_generation,
-            ):
-                self._library_skill_mutation_in_flight = False
-                if self.is_mounted:
-                    self._sync_library_skill_lifecycle_actions()
+    async def _run_library_skill_delete(self, skill_name: str, request_generation: int) -> None:
+        return await self._skills_controller._run_library_skill_delete(skill_name, request_generation)
 
     @on(Button.Pressed, "#library-skill-delete-cancel")
     def handle_library_skill_delete_cancel(self, event: Button.Pressed) -> None:
@@ -28314,99 +26934,11 @@ class LibraryScreen(BaseAppScreen):
             except (NoMatches, QueryError):
                 pass
 
-    async def _delete_library_skill(
-        self, skill_name: str, *, request_generation: int
-    ) -> None:
-        """Delete the selected Library skill, then return to the list view.
+    async def _delete_library_skill(self, skill_name: str, *, request_generation: int) -> None:
+        return await self._skills_controller._delete_library_skill(skill_name, request_generation=request_generation)
 
-        Args:
-            skill_name: The Library skill name to delete.
-        """
-        service = getattr(self.app_instance, "skills_scope_service", None)
-        delete_skill = getattr(service, "delete_skill", None)
-        if not callable(delete_skill):
-            self._update_library_skill_status_static("Skill deletion is unavailable.")
-            return
-        state = self._library_skill_editor_state
-        version = state.version if state is not None else None
-        try:
-            result = await self._run_library_service_call(
-                delete_skill,
-                skill_name,
-                mode="local",
-                expected_version=version,
-                isolate_in_worker=True,
-            )
-        except Exception:
-            logger.opt(exception=True).warning(
-                f"Failed to delete Library skill {skill_name!r}."
-            )
-            if not self._library_skill_detail_request_is_current(
-                skill_name=skill_name,
-                generation=request_generation,
-            ):
-                return
-            self._update_library_skill_status_static("Could not delete this skill.")
-            return
-
-        if not self._library_skill_detail_request_is_current(
-            skill_name=skill_name,
-            generation=request_generation,
-        ):
-            return
-
-        deleted = (
-            bool(result.get("deleted", True))
-            if isinstance(result, Mapping)
-            else bool(result)
-        )
-        if not deleted:
-            self._update_library_skill_status_static(
-                "This skill changed elsewhere — refresh and try again."
-            )
-            return
-
-        self._reset_library_skill_editor_state()
-        self._library_skills_filter = ""
-        self._refresh_library_skills_after_committed_mutation(
-            scope=dataclasses.replace(
-                self._library_skills_browse_controller.mutation_refresh_scope,
-                query="",
-            )
-        )
-        self._refresh_local_source_snapshot()
-
-    async def _request_library_skill_trust_passphrase(
-        self,
-        *,
-        title: str | None = None,
-        message: str | None = None,
-    ) -> str | None:
-        """Push the shared ``SkillTrustPassphraseModal`` and await a passphrase.
-
-        Mirrors ``skills_screen.SkillsScreen._request_skill_trust_passphrase``
-        (reused, not forked): never bootstraps from this editor, so
-        ``confirm_bootstrap`` is always ``False``. task-418: ``title``/
-        ``message`` let a caller present its real purpose (the approve
-        flow) instead of "Unlock Local Skill Trust".
-        """
-        push_screen_wait = getattr(self.app, "push_screen_wait", None)
-        if not callable(push_screen_wait):
-            notify = getattr(self.app_instance, "notify", None)
-            if callable(notify):
-                notify(
-                    "Local skill trust passphrase prompt is unavailable.",
-                    severity="warning",
-                )
-            return None
-        result = await push_screen_wait(
-            SkillTrustPassphraseModal(
-                confirm_bootstrap=False, title=title, message=message
-            )
-        )
-        if isinstance(result, str) and result:
-            return result
-        return None
+    async def _request_library_skill_trust_passphrase(self, *, title: str | None = None, message: str | None = None) -> str | None:
+        return await self._skills_controller._request_library_skill_trust_passphrase(title=title, message=message)
 
     async def _call_library_skill_trust_service(
         self,
@@ -28462,150 +26994,16 @@ class LibraryScreen(BaseAppScreen):
         return result, True
 
     async def _refresh_library_skill_trust_status(self) -> None:
-        """Re-fetch the open skill's trust status and patch the trust panel
-        in place (no recompose -- see ``_render_library_skill_trust_panel``).
-        """
-        name = self._selected_skill_name
-        generation = self._library_skill_detail_generation
-        state = self._library_skill_editor_state
-        if not name or state is None:
-            return
-        result, ok = await self._call_library_skill_trust_service(
-            "status_for_skill", name
-        )
-        if not ok or result is None:
-            return
-        if not self._library_skill_detail_request_is_current(
-            skill_name=name,
-            generation=generation,
-        ):
-            return
-        self._library_skill_editor_state = dataclasses.replace(
-            self._library_skill_editor_state,
-            trust_status=result.trust_status,
-            trust_blocked=result.trust_blocked,
-            trust_changed_files=tuple(result.changed_files),
-        )
-        self._render_library_skill_trust_panel()
-        self._update_library_skill_warnings_static(
-            name=self._read_library_skill_live_name()
-        )
+        return await self._skills_controller._refresh_library_skill_trust_status()
 
     def _refresh_library_skill_script_grant(self) -> None:
-        """Kick an off-thread read of the open skill's script-execution grant.
+        return self._skills_controller._refresh_library_skill_script_grant()
 
-        Task 7 (skills-script-execution): mirrors
-        ``_refresh_library_skills_trust_posture``'s shape rather than
-        routing through ``_call_library_skill_trust_service`` -- that
-        helper toasts a warning whenever the trust service is unavailable,
-        which is correct for an explicit user action (Unlock/Review/
-        Approve) but would be noisy fired silently on every skill open in
-        a deployment with no local trust service wired (e.g. server mode).
-        ``SkillTrustService.script_execution_granted`` re-scans the skill's
-        on-disk directory to verify its fingerprint, so -- like
-        ``trust_posture()`` -- it is NEVER called on the compose/event-loop
-        thread, only from here via ``asyncio.to_thread`` in
-        ``_load_library_skill_script_grant`` below. Called whenever the
-        skill editor opens on an existing skill
-        (``_apply_library_skill_detail``) and after a Save
-        (``_apply_library_skill_save_success``), since editing content
-        invalidates any prior grant.
-        """
-        name = self._selected_skill_name
-        generation = self._library_skill_detail_generation
-        trust_service = getattr(self.app_instance, "local_skill_trust_service", None)
-        granted_fn = getattr(trust_service, "script_execution_granted", None)
-        if not name or not callable(granted_fn):
-            self._library_skill_script_grant = False
-            return
-        self.run_worker(
-            self._load_library_skill_script_grant(name, generation, granted_fn),
-            exclusive=True,
-            group="library_skill_script_grant",
-            exit_on_error=False,
-        )
-
-    async def _load_library_skill_script_grant(
-        self,
-        name: str,
-        generation: int,
-        granted_fn,
-    ) -> None:
-        """Await the off-thread grant lookup and patch the trust panel.
-
-        Args:
-            name: The skill name captured at kick-off time (by
-                ``_refresh_library_skill_script_grant``), used to discard an
-                out-of-order result if a different skill is open by the
-                time this resolves.
-            generation: The retained Work session captured at kick-off time.
-            granted_fn: The trust service's bound
-                ``script_execution_granted`` method (captured at kick-off
-                so this never re-reads ``local_skill_trust_service``
-                itself).
-        """
-        try:
-            granted = await asyncio.to_thread(granted_fn, name)
-        except Exception:
-            granted = False
-        if not self._library_skill_detail_request_is_current(
-            skill_name=name,
-            generation=generation,
-        ):
-            return
-        self._library_skill_script_grant = bool(granted)
-
-        # task-8 (skills-script-execution) fix: NOT a direct call. This
-        # coroutine's own ``asyncio.to_thread`` round trip can resolve
-        # before ``_apply_library_skill_detail``'s own ``refresh(recompose=
-        # True)`` (posted moments earlier, on this same screen's message
-        # queue) has actually remounted the editor -- a real trust service
-        # doing real disk I/O usually loses that race, but there is no
-        # guarantee, and a fast trust service (or a slow recompose under
-        # load) can win it. A direct call here would then query widgets
-        # that do not exist YET, silently no-op through this method's own
-        # ``except (NoMatches, QueryError): pass`` guards, and never retry
-        # -- leaving the panel stuck showing "not granted"/disabled forever
-        # even though ``_library_skill_script_grant`` is correctly True in
-        # memory. ``call_after_refresh`` was the original answer, and
-        # task-15457 quietly invalidated it: the editor recompose it was
-        # ordering against became CANVAS-scoped (`_sync_library_canvas`,
-        # driven by the canvas's own message pump), which a SCREEN-level
-        # ``call_after_refresh`` has no ordering against -- the render fired
-        # before the canvas's children existed, swallowed ``NoMatches``, and
-        # never retried (task-15790, measured: grant stored True, render ran,
-        # button absent). Ride the same canvas post-recompose hook the
-        # caller's arming follow-up already rides; it runs `then` only once
-        # the canvas's new children are actually mounted.
-        def _render_then_arm() -> None:
-            self._render_library_skill_trust_panel()
-            if not self._library_skill_editor_armed:
-                self._arm_library_skill_editor()
-
-        _sync_library_canvas(self, "skills", then=_render_then_arm)
+    async def _load_library_skill_script_grant(self, name: str, generation: int, granted_fn) -> None:
+        return await self._skills_controller._load_library_skill_script_grant(name, generation, granted_fn)
 
     async def _request_library_skill_trust_bootstrap_passphrase(self) -> str | None:
-        """Push the confirm-passphrase bootstrap modal and await a passphrase.
-
-        Structural twin of ``_request_library_skill_trust_passphrase``: the
-        only difference is which modal it drives -- this one CREATES a
-        brand-new passphrase (twice-entry confirmed by
-        ``SkillTrustBootstrapModal`` itself), the other unlocks an existing
-        one.
-        """
-        push_screen_wait = getattr(self.app, "push_screen_wait", None)
-        if not callable(push_screen_wait):
-            notify = getattr(self.app_instance, "notify", None)
-            if callable(notify):
-                notify(
-                    "Local skill trust passphrase prompt is unavailable.",
-                    severity="warning",
-                )
-            return None
-        result = await push_screen_wait(SkillTrustBootstrapModal())
-        if isinstance(result, str) and result:
-            return result
-        return None
+        return await self._skills_controller._request_library_skill_trust_bootstrap_passphrase()
 
     @on(Button.Pressed, "#library-skills-trust-action")
     def handle_library_skills_trust_action(self, event: Button.Pressed) -> None:
@@ -28637,41 +27035,10 @@ class LibraryScreen(BaseAppScreen):
             self._open_first_blocked_skill()
 
     def _begin_library_skill_trust_setup(self) -> None:
-        self.run_worker(
-            self._setup_library_skill_trust(),
-            exclusive=True,
-            group="library_skill_trust",
-        )
+        return self._skills_controller._begin_library_skill_trust_setup()
 
     async def _setup_library_skill_trust(self) -> None:
-        """Set up trust from the list header's "setup"/"resetup" action.
-
-        Task 5 ambiguity resolution: "Set up" is reset-then-bootstrap ONLY
-        when a stale manifest already exists (``trust_store.has_manifest()``
-        -- the ``needs_resetup``/orphaned-manifest upgrade case). A truly
-        fresh install (no manifest at all, ``needs_setup``) skips the reset
-        entirely -- there is nothing to clear -- and goes straight to
-        ``bootstrap_trust``, matching the editor's own first-run
-        ``_bootstrap_library_skill_trust`` flow. Both the header's "setup"
-        and "resetup" action ids route here (``handle_library_skills_trust_action``
-        above); the manifest check itself is what decides whether a reset
-        actually happens, not the action id.
-        """
-        service = getattr(self.app_instance, "local_skill_trust_service", None)
-        if service is None:
-            return
-        passphrase = await self._request_library_skill_trust_bootstrap_passphrase()
-        if passphrase is None:
-            return
-        if getattr(service, "trust_store", None) and service.trust_store.has_manifest():
-            await self._call_library_skill_trust_service("reset_trust")
-        _, ok = await self._call_library_skill_trust_service(
-            "bootstrap_trust", passphrase
-        )
-        if ok:
-            self._refresh_library_skills_after_committed_mutation()
-            self._refresh_library_skills_trust_posture()
-            self._refresh_local_source_snapshot()
+        return await self._skills_controller._setup_library_skill_trust()
 
     @on(Button.Pressed, "#library-skills-trust-reset")
     def handle_library_skills_trust_reset_request(self, event: Button.Pressed) -> None:
@@ -28692,75 +27059,14 @@ class LibraryScreen(BaseAppScreen):
 
     @on(Button.Pressed, "#library-skills-trust-reset-cancel")
     def handle_library_skills_trust_reset_cancel(self, event: Button.Pressed) -> None:
-        """Back out of the Reset confirm row without resetting anything.
-
-        Args:
-            event: Button press event emitted by the confirm row's Cancel.
-        """
-        event.stop()
-        self._library_skill_trust_confirming_reset = False
-        if self.is_mounted:
-            _sync_library_canvas(self, "skills")
+        return self._skills_controller.handle_library_skills_trust_reset_cancel(event)
 
     @on(Button.Pressed, "#library-skills-trust-reset-confirm")
     def handle_library_skills_trust_reset_confirm(self, event: Button.Pressed) -> None:
-        """Run the confirmed destructive Reset (Task 5).
-
-        Args:
-            event: Button press event emitted by the confirm row's Reset.
-        """
-        event.stop()
-        self._library_skill_trust_confirming_reset = False
-        self.run_worker(
-            self._do_library_skill_trust_reset(),
-            exclusive=True,
-            group="library_skill_trust",
-        )
+        return self._skills_controller.handle_library_skills_trust_reset_confirm(event)
 
     async def _do_library_skill_trust_reset(self) -> None:
-        """Wipe all local trust state, then refresh whichever view is open.
-
-        Always refreshes the list header's posture (a stale ``locked``/
-        ``needs_resetup`` header must never linger after a successful
-        reset) and the shared local-source snapshot. When the editor's own
-        ``quarantined_manifest_error`` trust panel is what triggered this
-        (rather than the list header), also re-fetches the OPEN skill's own
-        trust status -- mirrors the per-skill re-fetch
-        ``_bootstrap_library_skill_trust`` already does after a bootstrap --
-        so the panel doesn't keep showing "manifest cannot be verified"
-        against a trust store that was just wiped.
-        """
-        _, ok = await self._call_library_skill_trust_service("reset_trust")
-        if not ok:
-            return
-        self._refresh_library_skills_after_committed_mutation()
-        self._refresh_library_skills_trust_posture()
-        self._refresh_local_source_snapshot()
-        name = self._selected_skill_name
-        if (
-            self._library_skills_view == "editor"
-            and name
-            and self._library_skill_editor_state is not None
-        ):
-            result, ok = await self._call_library_skill_trust_service(
-                "status_for_skill", name
-            )
-            if (
-                ok
-                and result is not None
-                and name == self._selected_skill_name
-                and self._library_skills_view == "editor"
-                and self._library_skill_editor_state is not None
-            ):
-                self._library_skill_editor_state = dataclasses.replace(
-                    self._library_skill_editor_state,
-                    trust_status=result.trust_status,
-                    trust_blocked=result.trust_blocked,
-                    trust_changed_files=tuple(result.changed_files),
-                )
-            self._library_skill_active_review = None
-        if self.is_mounted:
-            _sync_library_canvas(self, "skills")
+        return await self._skills_controller._do_library_skill_trust_reset()
 
     def _open_first_blocked_skill(self) -> None:
         """Open the first trust-blocked skill's editor (header "review" action).
@@ -28779,193 +27085,21 @@ class LibraryScreen(BaseAppScreen):
         )
 
     async def _open_library_skill_editor_for_review(self, skill_name: str) -> None:
-        """Open ``skill_name``'s editor, the same steps a real row press runs.
-
-        Deliberately NOT a refactor of ``handle_library_skill_row`` itself
-        (that handler stays untouched) -- just the same flush-veto, reset,
-        select, switch-to-editor-view, kick-the-off-thread-detail-fetch,
-        recompose sequence, sourced from a name instead of a row Button's
-        ``skill_name`` attribute.
-
-        Args:
-            skill_name: The blocked skill's name to open.
-        """
-        if not await self._flush_library_skill_save():
-            self._notify_skill_dirty_veto()
-            return
-        self._reset_library_skill_editor_state()
-        self._selected_skill_name = skill_name
-        self._library_selected_row_id = LIBRARY_ROW_BROWSE_SKILLS
-        self._library_skills_view = "editor"
-        self._library_skill_reader_mode = "trust"
-        self.run_worker(
-            self._refresh_library_skill_detail(skill_name),
-            exclusive=True,
-            group="library_skill_detail",
-        )
-        _sync_library_canvas(self, "skills")
+        return await self._skills_controller._open_library_skill_editor_for_review(skill_name)
 
     @on(Button.Pressed, "#library-skill-trust-setup")
     def handle_library_skill_trust_setup(self, event: Button.Pressed) -> None:
-        """Bootstrap local skill trust from the editor's first-run setup state.
-
-        Only rendered while ``trust_status == "trust_uninitialized"`` (a
-        brand-new, never-bootstrapped trust store) -- the Phase-1 gate fix
-        for the finding that a fresh install had no live-UI path to create
-        the trust passphrase at all.
-
-        Args:
-            event: Button press event emitted by the trust panel's "Set up
-                skill trust" action.
-        """
-        event.stop()
-        # task-417: any trust action supersedes a lingering "Saved.".
-        self._update_library_skill_status_static("")
-        self.run_worker(
-            self._bootstrap_library_skill_trust(),
-            exclusive=True,
-            group="library_skill_trust",
-        )
+        return self._skills_controller.handle_library_skill_trust_setup(event)
 
     async def _bootstrap_library_skill_trust(self) -> None:
-        """Create the initial trust baseline via a confirm-passphrase modal.
-
-        Unlike every other trust action here, ``bootstrap_trust`` is called
-        directly (never preceded by ``unlock_with_passphrase`` -- it takes
-        the new passphrase itself and derives+stores fresh keys). A full
-        recompose follows a successful bootstrap, not the usual targeted
-        ``_render_library_skill_trust_panel`` patch: the panel's layout
-        itself changes shape here, from the first-run setup state to the
-        normal Unlock/Review/Approve row, which a no-recompose patch can't
-        produce since those buttons don't exist in the DOM yet.
-        """
-        if (
-            self._library_skills_view != "editor"
-            or self._library_skill_editor_state is None
-        ):
-            return
-        name = self._selected_skill_name
-        generation = self._library_skill_detail_generation
-        passphrase = await self._request_library_skill_trust_bootstrap_passphrase()
-        if passphrase is None:
-            return
-        _, ok = await self._call_library_skill_trust_service(
-            "bootstrap_trust", passphrase
-        )
-        if not ok:
-            return
-        self._refresh_library_skills_after_committed_mutation()
-        if (
-            name
-            and self._library_skill_detail_request_is_current(
-                skill_name=name,
-                generation=generation,
-            )
-            and self._library_skill_editor_state is not None
-        ):
-            result, status_ok = await self._call_library_skill_trust_service(
-                "status_for_skill", name
-            )
-            if (
-                status_ok
-                and result is not None
-                and self._library_skill_detail_request_is_current(
-                    skill_name=name,
-                    generation=generation,
-                )
-            ):
-                self._library_skill_editor_state = dataclasses.replace(
-                    self._library_skill_editor_state,
-                    trust_status=result.trust_status,
-                    trust_blocked=result.trust_blocked,
-                    trust_changed_files=tuple(result.changed_files),
-                )
-        self._refresh_local_source_snapshot()
-        if name and self._library_skill_detail_request_is_current(
-            skill_name=name,
-            generation=generation,
-        ):
-            self._library_skill_active_review = None
-        if (
-            name
-            and self.is_mounted
-            and self._library_skill_detail_request_is_current(
-                skill_name=name,
-                generation=generation,
-            )
-        ):
-            # Disarm dirty-tracking before the recompose (mirrors
-            # ``_apply_library_skill_detail``): remounting the Inputs with
-            # their existing values still fires their initial
-            # ``Input.Changed`` -- without this, still-armed dirty-tracking
-            # would misread that as a real edit and wrongly mark the editor
-            # dirty (vetoing the next Back/row-switch for no reason).
-            self._library_skill_dirty = False
-            self._library_skill_editor_armed = False
-            _sync_library_canvas(
-                self,
-                "skills",
-                then=self._arm_library_skill_editor,
-            )
+        return await self._skills_controller._bootstrap_library_skill_trust()
 
     @on(Button.Pressed, "#library-skill-trust-unlock")
     def handle_library_skill_trust_unlock(self, event: Button.Pressed) -> None:
-        """Unlock local skill trust for this session via the passphrase modal.
-
-        Args:
-            event: Button press event emitted by the trust panel's "Unlock"
-                action.
-        """
-        event.stop()
-        # task-417: any trust action supersedes a lingering "Saved.".
-        self._update_library_skill_status_static("")
-        self.run_worker(
-            self._unlock_library_skill_trust(),
-            exclusive=True,
-            group="library_skill_trust",
-        )
+        return self._skills_controller.handle_library_skill_trust_unlock(event)
 
     async def _unlock_library_skill_trust(self) -> None:
-        """Unlock local skill trust for this session via the passphrase modal.
-
-        Task 5: reused by BOTH the editor's own "Unlock" trust-panel action
-        AND the list header's "unlock" action (posture ``locked``) -- this
-        originally returned immediately unless already in the editor,
-        which made the header's Unlock a silent no-op (browsing the list
-        is exactly where ``locked`` posture is shown). Editor mode still
-        gets the original no-recompose targeted panel patch
-        (``_refresh_library_skill_trust_status``) so an in-progress unsaved
-        edit elsewhere in the editor is never discarded by a full rebuild;
-        list mode instead refreshes the header's posture AND the shared
-        local-source snapshot -- the list rows' trust glyphs and the
-        header's blocked-count both derive from that snapshot
-        (``_build_library_skills_state``), so refreshing posture alone
-        would leave every row's ``⚠``/``✓`` and the "N need review" count
-        stale until some later snapshot refresh (matches the sibling
-        ``_setup_library_skill_trust``/``_do_library_skill_trust_reset``
-        handlers). The list view has no unsaved-edit state to lose.
-        """
-        editor_name = (
-            self._selected_skill_name if self._library_skills_view == "editor" else None
-        )
-        generation = self._library_skill_detail_generation
-        passphrase = await self._request_library_skill_trust_passphrase()
-        if passphrase is None:
-            return
-        _, ok = await self._call_library_skill_trust_service(
-            "unlock_with_passphrase", passphrase
-        )
-        if not ok:
-            return
-        self._refresh_library_skills_after_committed_mutation()
-        if editor_name and self._library_skill_detail_request_is_current(
-            skill_name=editor_name,
-            generation=generation,
-        ):
-            await self._refresh_library_skill_trust_status()
-        elif editor_name is None and self._library_skills_view != "editor":
-            self._refresh_library_skills_trust_posture()
-            self._refresh_local_source_snapshot()
+        return await self._skills_controller._unlock_library_skill_trust()
 
     @on(Button.Pressed, "#library-skill-trust-review")
     def handle_library_skill_trust_review(self, event: Button.Pressed) -> None:
@@ -28985,39 +27119,11 @@ class LibraryScreen(BaseAppScreen):
         )
 
     async def _review_library_skill_trust(self) -> None:
-        if self._library_skills_view != "editor" or not self._selected_skill_name:
-            return
-        name = self._selected_skill_name
-        generation = self._library_skill_detail_generation
-        result, ok = await self._call_library_skill_trust_service(
-            "capture_review", name
-        )
-        if not ok or not isinstance(result, Mapping) or not result.get("review_id"):
-            return
-        if not self._library_skill_detail_request_is_current(
-            skill_name=name,
-            generation=generation,
-        ):
-            return
-        self._library_skill_active_review = dict(result)
-        self._render_library_skill_trust_panel()
+        return await self._skills_controller._review_library_skill_trust()
 
     @on(Button.Pressed, "#library-skill-trust-approve")
     def handle_library_skill_trust_approve(self, event: Button.Pressed) -> None:
-        """Approve the captured trust review via the passphrase modal.
-
-        Args:
-            event: Button press event emitted by the trust panel's
-                "Approve" action.
-        """
-        event.stop()
-        # task-417: any trust action supersedes a lingering "Saved.".
-        self._update_library_skill_status_static("")
-        self.run_worker(
-            self._approve_library_skill_trust(),
-            exclusive=True,
-            group="library_skill_trust",
-        )
+        return self._skills_controller.handle_library_skill_trust_approve(event)
 
     async def _approve_library_skill_trust(self) -> None:
         if (
@@ -29086,48 +27192,10 @@ class LibraryScreen(BaseAppScreen):
 
     @on(Button.Pressed, "#library-skill-script-grant-revoke")
     def handle_library_skill_script_grant_revoke(self, event: Button.Pressed) -> None:
-        """Revoke the open skill's standing script-execution grant.
-
-        Args:
-            event: Button press event emitted by the trust panel's "Revoke
-                script access" action.
-        """
-        event.stop()
-        self.run_worker(
-            self._revoke_library_skill_script_grant(),
-            exclusive=True,
-            group="library_skill_trust",
-        )
+        return self._skills_controller.handle_library_skill_script_grant_revoke(event)
 
     async def _revoke_library_skill_script_grant(self) -> None:
-        """Drop the open skill's standing script grant, then patch the panel.
-
-        Task 7 (skills-script-execution): the counterpart to whatever
-        confirm-card action (Task 6) called ``grant_script_execution`` in
-        the first place -- a grant the user cannot see or withdraw here
-        would be a real hole. ``revoke_script_execution`` raises
-        ``ValueError`` on a malformed skill name, which
-        ``_call_library_skill_trust_service`` already catches and reports
-        via the standard failure toast, so no extra guard is needed here.
-        The outcome is applied directly (no re-fetch) since a successful
-        revoke unambiguously means "not granted".
-        """
-        if self._library_skills_view != "editor" or not self._selected_skill_name:
-            return
-        name = self._selected_skill_name
-        generation = self._library_skill_detail_generation
-        _, ok = await self._call_library_skill_trust_service(
-            "revoke_script_execution", name
-        )
-        if not ok:
-            return
-        if not self._library_skill_detail_request_is_current(
-            skill_name=name,
-            generation=generation,
-        ):
-            return
-        self._library_skill_script_grant = False
-        self._render_library_skill_trust_panel()
+        return await self._skills_controller._revoke_library_skill_script_grant()
 
     @on(Button.Pressed, "#library-prompts-import")
     async def handle_library_prompts_import(self, event: Button.Pressed) -> None:
